bengali slogan added in homepage
