+++ src/pages/HomePage/Mask/Narrowscreen/bn/index.tsx
@@ -1,0 +1,21 @@
+const Bn = () => {
+  return (
+    <g>
+      <g id="mask">
+        <path className="masking__layer" d="M187.25952,490.11493V487.2016c-1.06567-.05579-2.04913-.13721-2.94409-.24628a26.54526,26.54526,0,0,1-4.29474-.86c.053.13049.132.24658.17761.38128a9.67527,9.67527,0,0,1,.46277,2.64173,9.527,9.527,0,0,1,1.83515.54187,12.84805,12.84805,0,0,1,3.42285,2.13025,19.37,19.37,0,0,1,1.42645,1.37359c-.02911-.40576-.05212-.7915-.06207-1.13415Q187.25873,491.21631,187.25952,490.11493Z" transform="translate(0.33691 -0.18781)" />
+        <path className="masking__layer" d="M209.07489,439.30469v4.83508a8.51659,8.51659,0,0,0,.16766,2.01062,1.8735,1.8735,0,0,0,.55048.95746,2.36605,2.36605,0,0,0,1.81909.62231,5.48424,5.48424,0,0,0,2.29791-.4787,9.66925,9.66925,0,0,0,2.10633-1.34039,7.056,7.056,0,0,0,1.55584-1.81915,3.815,3.815,0,0,0,.5506-1.91492,1.50314,1.50314,0,0,0-.4787-1.24463,2.2716,2.2716,0,0,0-1.43615-.383h-.383a9.39593,9.39593,0,0,1-2.5133-.31122,10.06416,10.06416,0,0,1-2.41755-1.07709,9.86311,9.86311,0,0,1-1.69952-1.24469c-.07391-.0672-.12659-.1405-.19824-.20831.01025.12982.02038.23785.0307.37591Q209.07428,438.73026,209.07489,439.30469Z" transform="translate(0.33691 -0.18781)" />
+        <path className="masking__layer" d="M149.56482,489.82764v-5.32947c-1.54266.44623-3.03.93024-4.42816,1.47577q-2.51322.98172-5.242,2.274a17.76665,17.76665,0,0,1,3.78186,1.22065,26.687,26.687,0,0,1,3.66224,2.01062,24.38179,24.38179,0,0,1,2.31945,1.75512c-.0169-.29052-.03357-.58184-.04559-.89337Q149.56461,491.12027,149.56482,489.82764Z" transform="translate(0.33691 -0.18781)" />
+        <path className="masking__layer" d="M187.25952,481.52368v-4.23852h-15.4798c.061.12817.13074.24926.18464.383a9.53122,9.53122,0,0,1,.56829,2.286c.21381.0296.3949.09345.59925.13306a62.72912,62.72912,0,0,0,7.16223,1.09955C182.49207,481.39673,184.82788,481.49786,187.25952,481.52368Z" transform="translate(0.33691 -0.18781)" />
+        <path className="masking__layer" d="M265.70776,443.99622v-6.12909c-1.52514.44244-2.99682.92163-4.38037,1.46148q-2.51329.98172-5.242,2.27393a16.3089,16.3089,0,0,1,3.80585,1.24463,24.03714,24.03714,0,0,1,3.87768,2.25006,16.65113,16.65113,0,0,1,2.06024,1.75366c-.01544-.22143-.0354-.42566-.04962-.65265Q265.70777,445.04944,265.70776,443.99622Z" transform="translate(0.33691 -0.18781)" />
+        <path className="masking__layer" d="M0,.18781V599.81219H500V.18781ZM75.31433,315.39337a74.32612,74.32612,0,0,1-9.02295-8.98291,7.90874,7.90874,0,0,1,0-10.14,74.32511,74.32511,0,0,1,9.02295-8.98279,97.97217,97.97217,0,0,1,27.069-16.47839,96.68034,96.68034,0,0,1,36.092-7.08985,93.93791,93.93791,0,0,1,36.092,6.83655,96.068,96.068,0,0,1,18.04614,9.63684,97.043,97.043,0,0,1,9.02283,6.95868,73.677,73.677,0,0,1,9.023,9.119,7.98911,7.98911,0,0,1,0,10.14,73.69462,73.69462,0,0,1-9.02289,9.119,97.06293,97.06293,0,0,1-9.023,6.95868,96.07173,96.07173,0,0,1-18.04614,9.6369,93.94141,93.94141,0,0,1-36.092,6.83649,96.68277,96.68277,0,0,1-36.092-7.08991A97.96859,97.96859,0,0,1,75.31433,315.39337Zm63.739,130.66132a4.75018,4.75018,0,0,0,2.03455,2.08239,5.45119,5.45119,0,0,0,2.41754.59839,9.50564,9.50564,0,0,0,1.6994-.14361,5.99409,5.99409,0,0,0,1.69953-.62226l2.34576,5.4574a11.684,11.684,0,0,1-3.04,1.05316,17.54818,17.54818,0,0,1-3.27917.28717,13.70742,13.70742,0,0,1-6.34308-1.38818,9.59606,9.59606,0,0,1-4.18884-4.23675,15.15877,15.15877,0,0,1-1.48407-7.0611,16.70351,16.70351,0,0,1,.93353-5.57709,18.75185,18.75185,0,0,1,2.56128-4.883,21.231,21.231,0,0,1,3.7818-3.97339,21.945,21.945,0,0,1,4.64368-2.89624h6.17547v5.88824h-2.92023a20.4925,20.4925,0,0,0-3.08776,2.15424,16.6865,16.6865,0,0,0-2.48932,2.609,11.461,11.461,0,0,0-1.67548,3.0639,10.09567,10.09567,0,0,0-.59851,3.4707A8.631,8.631,0,0,0,139.05328,446.05469Zm129.76966,31.23047H252.25916v6.94146a2.42728,2.42728,0,0,0,.2633,1.22071,1.03411,1.03411,0,0,0,.98139.45483,2.49447,2.49447,0,0,0,1.2926-.407,5.68807,5.68807,0,0,0,1.3164-1.14887,20.435,20.435,0,0,0,1.31653-1.6994l5.88831-1.05328a18.75139,18.75139,0,0,1,2.58508,4.57184,13.33282,13.33282,0,0,1,.8617,4.76324,9.45757,9.45757,0,0,1-1.43616,5.02655,10.1702,10.1702,0,0,1-4.40424,3.78192,17.63537,17.63537,0,0,1-7.6596,1.43616,22.78744,22.78744,0,0,1-7.06116-1.07709,18.01428,18.01428,0,0,1-6.12757-3.44677,22.40336,22.40336,0,0,1-5.00262-6.34308l5.84039-3.2074a25.38145,25.38145,0,0,0,3.27918,4.35626,11.797,11.797,0,0,0,3.758,2.68091,11.36883,11.36883,0,0,0,4.64362.90961,8.85845,8.85845,0,0,0,5.14618-1.29261,4.28029,4.28029,0,0,0,1.84314-3.73407,5.597,5.597,0,0,0-.11926-.96923,13.21692,13.21692,0,0,1-3.47113,1.99859,11.0346,11.0346,0,0,1-3.87769.69409,9.28679,9.28679,0,0,1-3.08771-.4787,5.42382,5.42382,0,0,1-2.274-1.48406,5.60632,5.60632,0,0,1-.40516-.57068,45.64705,45.64705,0,0,0-5.19586-3.307,11.63413,11.63413,0,0,0-5.4574-1.53192,2.45474,2.45474,0,0,0-1.55584.40692,1.22489,1.22489,0,0,0-.50269.98138,1.43873,1.43873,0,0,0,.38306,1.02924,3.80814,3.80814,0,0,0,1.34033.78992l-3.11169,5.45746a11.77167,11.77167,0,0,1-4.42817-3.1836,6.62349,6.62349,0,0,1-1.60369-4.33246,6.26662,6.26662,0,0,1,1.19677-3.97338,6.76254,6.76254,0,0,1,3.18354-2.226,12.76,12.76,0,0,1,4.23663-.69421,16.19469,16.19469,0,0,1,4.11706.52667,17.9261,17.9261,0,0,1,3.99732,1.6037,20.01069,20.01069,0,0,1,2.13324,1.35687c-.00818-.17767-.021-.32818-.02686-.51916q-.02481-.76511-.02392-1.4361v-2.87237H221.095v23.88818h-7.27661V480.30115a5.31818,5.31818,0,0,0-1.55579-2.22614,3.88291,3.88291,0,0,0-2.5133-.78985H194.48816V481.229a30.51813,30.51813,0,0,0,3.39892-.56878,41.74376,41.74376,0,0,0,4.45215-1.31653l5.21808,4.21271a15.1917,15.1917,0,0,0-1.41223,3.99737,20.46229,20.46229,0,0,0-.45477,4.18878,24.52593,24.52593,0,0,0,.28717,3.73407,24.048,24.048,0,0,0,.9096,3.78186l-7.18078,1.38831a30.29231,30.29231,0,0,1-.98139-4.47607,31.87235,31.87235,0,0,1-.31121-4.42817,21.5611,21.5611,0,0,1,.52661-4.83508c.0376-.16431.09936-.31311.141-.47479a24.41451,24.41451,0,0,1-3.516.64239c-.34271.03467-.71869.05164-1.07715.078v14.02032h-7.22864v-.36993a33.0677,33.0677,0,0,0-2.44147-3.1247,12.536,12.536,0,0,0-2.5133-2.20209,4.77712,4.77712,0,0,0-2.60907-.81384,2.4245,2.4245,0,0,0-1.02924.19141,1.39925,1.39925,0,0,0-.62231.55053,1.79677,1.79677,0,0,0-.21539.93354,4.52089,4.52089,0,0,0,.11962,1.00531,8.27147,8.27147,0,0,0,.35907,1.149l-7.08514,1.05317q-.3837-1.0524-.64624-2.03455a7.8025,7.8025,0,0,1-.26331-2.03455,5.88507,5.88507,0,0,1,.74207-3.0639,5.99437,5.99437,0,0,1,1.93884-2.01062,8.48451,8.48451,0,0,1,.86023-.45252c.00012-.02612.00147-.04772.00147-.07409a7.27049,7.27049,0,0,0-.26337-1.819,4.61029,4.61029,0,0,0-.86169-1.74738,1.95565,1.95565,0,0,0-1.6037-.74206,1.52156,1.52156,0,0,0-1.12494.40692,1.58772,1.58772,0,0,0-.407,1.17291,4.833,4.833,0,0,0,.09576.95746,3.39283,3.39283,0,0,0,.28723.86163l-7.08514,1.05316a11.2138,11.2138,0,0,1-.59839-1.843,8.17857,8.17857,0,0,1-.21533-1.84308,5.787,5.787,0,0,1,.74194-3.01593,5.919,5.919,0,0,1,1.96277-2.01062,8.60577,8.60577,0,0,1,.80469-.41345,5.50983,5.50983,0,0,0-.92432-2.88978,5.20526,5.20526,0,0,0-.63562-.71807h-7.23468v29.48919h-7.18091a4.65287,4.65287,0,0,0-1.72333-1.74725,6.06272,6.06272,0,0,0-2.96808-.59839,16.82029,16.82029,0,0,0-2.34576.16754q-1.19715.16709-2.51324.35907a20.289,20.289,0,0,1-2.89631.19141,16.22364,16.22364,0,0,1-3.0877-.31116,8.54952,8.54952,0,0,1-2.992-1.19678,6.47857,6.47857,0,0,1-2.24994-2.53723,9.25212,9.25212,0,0,1-.86175-4.2846q0-.86162.07178-1.65155a8.99169,8.99169,0,0,1,.26324-1.50794l6.94147.76587q-.09641.43093-.16748.88568a5.98922,5.98922,0,0,0-.07184.93353,2.60062,2.60062,0,0,0,1.00531,2.36969,4.60632,4.60632,0,0,0,2.53723.64624,12.486,12.486,0,0,0,1.98669-.14361q.88531-.14355,1.77131-.28729a11.82,11.82,0,0,1,1.891-.14356,5.93305,5.93305,0,0,1,2.89618.67017,13.78665,13.78665,0,0,1,2.46539,1.7713V500.168a36.58337,36.58337,0,0,0-4.52387-3.63824,26.31439,26.31439,0,0,0-4.97869-2.60907,34.099,34.099,0,0,0-6.24726-1.74732l-2.68091-6.51056a100.06326,100.06326,0,0,1,9.52655-4.38031,78.04259,78.04259,0,0,1,9.14362-2.99768v-.99969H126.2511v-5.88825h84.16833a3.66361,3.66361,0,0,1,1.96277.59839,4.58882,4.58882,0,0,1,1.62763,2.27393,8.675,8.675,0,0,1-.14356-1.31647q-.04833-.88522-.04791-1.508v-3.30322h6.07977l1.149,3.25537h47.77582Zm-17.90772-25.43793a3.49311,3.49311,0,0,1,1.05316-2.60907,3.73876,3.73876,0,0,1,2.7287-1.0293,3.6796,3.6796,0,0,1,2.68091,1.0293,3.49585,3.49585,0,0,1,1.05316,2.60907,3.55311,3.55311,0,0,1-1.05316,2.65686,3.67907,3.67907,0,0,1-2.68091,1.02924,3.73824,3.73824,0,0,1-2.7287-1.02924A3.55036,3.55036,0,0,1,250.91522,451.84723Zm26.42542-21.2074h-4.35639V454.528h-7.27649a20.88847,20.88847,0,0,0-4.165-4.21277,24.05647,24.05647,0,0,0-5.1941-2.968,29.91133,29.91133,0,0,0-6.34307-1.81921l-2.68079-6.51056a100.06907,100.06907,0,0,1,9.52655-4.38037,79.12562,79.12562,0,0,1,9.14355-3.00977v-.98755H243.92578a21.6,21.6,0,0,0-3.11175,2.15424,16.16335,16.16335,0,0,0-2.51325,2.609,11.46312,11.46312,0,0,0-1.67554,3.0639,10.1,10.1,0,0,0-.59832,3.4707,8.63257,8.63257,0,0,0,.81372,4.117,4.75173,4.75173,0,0,0,2.03454,2.08239,5.45166,5.45166,0,0,0,2.41761.59839,9.50757,9.50757,0,0,0,1.69952-.14361,5.99405,5.99405,0,0,0,1.6994-.62226l2.34577,5.4574a11.684,11.684,0,0,1-3.03986,1.05316,17.5496,17.5496,0,0,1-3.2793.28717,13.74685,13.74685,0,0,1-6.31915-1.38818,9.56672,9.56672,0,0,1-4.21271-4.23675,15.15816,15.15816,0,0,1-1.484-7.0611,14.60428,14.60428,0,0,1,1.14887-5.76862,14.81581,14.81581,0,0,1,3.25537-4.78717,14.57574,14.57574,0,0,1,1.11084-.9909c-.28247.02289-.57977.04223-.89545.05737q-.98181.04815-2.56115.04785H208.97919v.37341a9.96224,9.96224,0,0,0,1.55584,1.63721,7.999,7.999,0,0,0,2.609,1.38837,10.71308,10.71308,0,0,0,3.30323.4787h1.14886a13.53953,13.53953,0,0,1,3.68622.40686,5.41165,5.41165,0,0,1,2.34576,1.36438,4.19675,4.19675,0,0,1,1.26856,1.867,7.42083,7.42083,0,0,1,.40692,2.48938,10.981,10.981,0,0,1-1.101,4.9787,14.68978,14.68978,0,0,1-2.92023,3.97339,15.73692,15.73692,0,0,1-4.52393,2.92017,13.96249,13.96249,0,0,1-5.577,1.10107,13.55041,13.55041,0,0,1-3.99744-.50269,7.89528,7.89528,0,0,1-2.70471-1.41223,7.687,7.687,0,0,1-1.93884-2.75262,13.68931,13.68931,0,0,1-.742-5.19416V430.63983H194.283V454.528h-7.27661v-20.8722a5.31815,5.31815,0,0,0-1.55579-2.22613,3.88291,3.88291,0,0,0-2.5133-.78986h-3.63794V454.528h-7.27661V436.45111c-.47235-.55835-.93653-1.11859-1.43616-1.67035a22.36812,22.36812,0,0,0-2.44147-2.32184,11.67112,11.67112,0,0,0-2.7287-1.67547,7.75071,7.75071,0,0,0-3.11175-.62232,6.04578,6.04578,0,0,0-3.56641.90961,19.08862,19.08862,0,0,0-1.49805,1.21179c.25745-.10577.51307-.21509.78-.30231a9.629,9.629,0,0,1,2.94416-.38293,7.35188,7.35188,0,0,1,3.49469.76593,5.563,5.563,0,0,1,2.226,2.10645,6.11085,6.11085,0,0,1,.79,3.15948,8.2283,8.2283,0,0,1-2.15436,5.40961,21.48661,21.48661,0,0,1-6.41486,4.78711l-4.02124-5.601a25.859,25.859,0,0,0,4.26068-2.36969,2.81808,2.81808,0,0,0,1.38831-2.1781,1.39678,1.39678,0,0,0-.43085-1.10107,2.06416,2.06416,0,0,0-1.38837-.383,6.24156,6.24156,0,0,0-1.84307.2873,11.69217,11.69217,0,0,0-2.0824.90948l-5.45746-4.45209a34.28631,34.28631,0,0,1,3.66217-4.38025,16.23334,16.23334,0,0,1,4.09314-3.01605,10.79082,10.79082,0,0,1,4.9787-1.12494,11.17777,11.17777,0,0,1,4.14093.766,15.84685,15.84685,0,0,1,3.87768,2.27387q.98611.767,2.04059,1.73987c-.00958-.19342-.0166-.36389-.03-.59088q-.07178-1.22078-.07183-2.70484v-4.40423h6.12762l1.101,3.25537h4.4043v.0072a3.66889,3.66889,0,0,1,1.91455.59119,4.58883,4.58883,0,0,1,1.62763,2.27392,8.67343,8.67343,0,0,1-.14356-1.31653q-.04835-.88512-.04791-1.50793v-3.30322h6.07977l1.149,3.25537h83.10547Zm3.54364-221.9516a96.07084,96.07084,0,0,1-9.6369,18.04615,97.04812,97.04812,0,0,1-6.95868,9.023,73.6906,73.6906,0,0,1-9.119,9.02282,7.98889,7.98889,0,0,1-10.14,0,73.68183,73.68183,0,0,1-9.119-9.023,97.03365,97.03365,0,0,1-6.95868-9.02288,96.06653,96.06653,0,0,1-9.63684-18.04615,93.93964,93.93964,0,0,1-6.83655-36.092,96.68171,96.68171,0,0,1,7.08984-36.092,97.97275,97.97275,0,0,1,16.4784-27.069,74.3261,74.3261,0,0,1,8.98291-9.02295,7.90872,7.90872,0,0,1,10.13989,0,74.32682,74.32682,0,0,1,8.98291,9.02295,97.97045,97.97045,0,0,1,16.47833,27.069,96.68171,96.68171,0,0,1,7.08984,36.092A93.9411,93.9411,0,0,1,280.88428,208.68823ZM373.9245,503.10632H366.6V473.32983h7.32446Zm59.98364-196.69586a74.32157,74.32157,0,0,1-9.023,8.98291,97.96844,97.96844,0,0,1-27.069,16.47827,96.68158,96.68158,0,0,1-36.092,7.08991,93.94111,93.94111,0,0,1-36.092-6.83649,96.06851,96.06851,0,0,1-18.0462-9.6369,97.055,97.055,0,0,1-9.02295-6.95868,73.69532,73.69532,0,0,1-9.02289-9.119,7.98911,7.98911,0,0,1,0-10.14,73.69026,73.69026,0,0,1,9.023-9.119,97.0359,97.0359,0,0,1,9.02289-6.95868,96.06478,96.06478,0,0,1,18.0462-9.63684,93.93627,93.93627,0,0,1,36.092-6.83655,96.68033,96.68033,0,0,1,36.092,7.08985,97.972,97.972,0,0,1,27.069,16.47839,74.32059,74.32059,0,0,1,9.023,8.98279A7.84941,7.84941,0,0,1,433.90814,306.41046Z" transform="translate(0.33691 -0.18781)" />
+      </g>
+      <g id="visible">
+        <path className="visible__layer" d="M351.88464,496.859c-.10576-.036-.21371-.06375-.31978-.099a11.92742,11.92742,0,0,0,3.76647-1.95953A7.01644,7.01644,0,0,0,357.701,489.032a7.91793,7.91793,0,0,0-1.19676-4.141,8.6929,8.6929,0,0,0-3.85373-3.25525,16.54857,16.54857,0,0,0-7.06111-1.26865c-.0974,0-.17828.01713-.27453.01828a5.34748,5.34748,0,0,0-.51541-.71239,5.83415,5.83415,0,0,0-.46524-.45479h16.57427v-5.88829h-4.5958a14.30982,14.30982,0,0,0-.64624-4.06911,6.43046,6.43046,0,0,0-1.60369-2.58518,7.02479,7.02479,0,0,0-2.89628-1.46,19.52659,19.52659,0,0,0-4.476-.40693h-2.72866a12.75342,12.75342,0,0,1-2.633-.19155,3.25093,3.25093,0,0,1-1.34034-.62231,2.79573,2.79573,0,0,1-.69421-1.17283,5.60126,5.60126,0,0,1-.21538-1.65166,4.11832,4.11832,0,0,1,.04786-.57445c.03146-.2229.06354-.47873.09573-.7659l-6.8936-.43086q-.1436.86167-.19145,1.508-.04719.64629-.04786,1.31652a11.64208,11.64208,0,0,0,.40693,3.25524,6.03345,6.03345,0,0,0,1.31642,2.39363,7.50671,7.50671,0,0,0,3.68621,2.13028,23.96489,23.96489,0,0,0,6.17546.64624h2.68079a3.34756,3.34756,0,0,1,2.32184.62242,2.91058,2.91058,0,0,1,.6942,2.05848H328.5475l-1.1489-3.25535h-6.07984v3.30321q0,.62278.048,1.508a8.6508,8.6508,0,0,0,.14358,1.31652,4.59042,4.59042,0,0,0-1.62762-2.274,3.66556,3.66556,0,0,0-1.94792-.59618v-.0022H281.21684v5.88829h36.03281a3.88269,3.88269,0,0,1,2.51328.78983,5.31691,5.31691,0,0,1,1.55583,2.2261v20.87225h7.2766V479.21815H335.837a6.44259,6.44259,0,0,1,2.21253,1.29248,4.92614,4.92614,0,0,1,.89088,1.12706,7.52457,7.52457,0,0,0-2.32706,1.60171,4.38809,4.38809,0,0,0-1.26866,3.11166,10.22949,10.22949,0,0,0,.16763,1.867,8.55622,8.55622,0,0,0,.50265,1.72335l6.84563-.86162q-.09641-.23905-.19144-.55052a2.01367,2.01367,0,0,1-.09573-.59848,1.66219,1.66219,0,0,1,.35907-1.101,2.1762,2.1762,0,0,1,1.101-.64624,6.926,6.926,0,0,1,1.891-.21548,5.32359,5.32359,0,0,1,3.3031.88565,2.84086,2.84086,0,0,1,1.1489,2.3697,3.02986,3.02986,0,0,1-.83769,2.20207,5.41659,5.41659,0,0,1-2.51328,1.31652,16.61244,16.61244,0,0,1-4.26066.45479,26.701,26.701,0,0,1-3.59039-.23931q-1.77162-.239-3.44679-.62241l-1.1489,5.60111a53.13212,53.13212,0,0,1,6.48666,1.31642,58.0963,58.0963,0,0,1,5.96008,1.96276,57.22143,57.22143,0,0,1,5.45743,2.46542,53.41937,53.41937,0,0,1,4.835,2.82448l3.68622-5.07442a34.32543,34.32543,0,0,0-4.5-2.657A39.37211,39.37211,0,0,0,351.88464,496.859Z" transform="translate(0.33691 -0.18781)" />
+        <path className="visible__layer" d="M310.82579,482.06657a13.25709,13.25709,0,0,0-6.29522-1.36439,13.76755,13.76755,0,0,0-4.33245.67018,7.04832,7.04832,0,0,0-3.27918,2.15421A5.98177,5.98177,0,0,0,295.67421,487.5a9.02737,9.02737,0,0,0,.19145,1.69952,10.0579,10.0579,0,0,0,.67028,2.08242l6.84573-.90959a7.5008,7.5008,0,0,1-.23941-.90958,5.076,5.076,0,0,1-.09572-1.00532,1.89455,1.89455,0,0,1,.62231-1.53189,2.36058,2.36058,0,0,1,1.57976-.52659,3.43877,3.43877,0,0,1,1.77131.47872,3.23233,3.23233,0,0,1,1.26866,1.41225,5.16385,5.16385,0,0,1,.45469,2.274,5.31465,5.31465,0,0,1-.43076,2.20207,3.65665,3.65665,0,0,1-1.31652,1.55583,6.29355,6.29355,0,0,1-2.15421.88566,13.43938,13.43938,0,0,1-2.94414.28727,10.019,10.019,0,0,1-5.43349-1.484,12.52324,12.52324,0,0,1-4.141-4.64367,37.3231,37.3231,0,0,1-3.15952-8.13822l-6.75,1.867a34.33,34.33,0,0,0,5.0506,11.10629,17.41072,17.41072,0,0,0,6.94146,5.76863,20.91894,20.91894,0,0,0,8.44943,1.69942,17.42212,17.42212,0,0,0,7.468-1.38831A9.65735,9.65735,0,0,0,314.65549,496.5a10.26518,10.26518,0,0,0,1.41224-5.31383,10.71757,10.71757,0,0,0-1.34035-5.40956A9.53418,9.53418,0,0,0,310.82579,482.06657Z" transform="translate(0.33691 -0.18781)" />
+        <path className="visible__layer" d="M360.14108,447.54645c-.10576-.036-.21371-.06375-.31978-.099a11.92892,11.92892,0,0,0,3.76658-1.95952,7.01673,7.01673,0,0,0,2.36959-5.76853,7.918,7.918,0,0,0-1.19676-4.141,8.693,8.693,0,0,0-3.85373-3.25525,16.54826,16.54826,0,0,0-7.06112-1.26866c-.09739,0-.17828.01714-.27452.01829a5.34748,5.34748,0,0,0-.51541-.71239,5.8342,5.8342,0,0,0-.46524-.4548H369.165v-5.88828h-4.5957a14.309,14.309,0,0,0-.64635-4.06911,6.43028,6.43028,0,0,0-1.60369-2.58518,7.02461,7.02461,0,0,0-2.89628-1.46,19.5265,19.5265,0,0,0-4.476-.40693h-2.72866a12.75342,12.75342,0,0,1-2.633-.19155,3.25092,3.25092,0,0,1-1.34035-.62232,2.79455,2.79455,0,0,1-.6942-1.17283,5.60114,5.60114,0,0,1-.21538-1.65165,4.11667,4.11667,0,0,1,.04786-.57445c.03135-.22291.06354-.47873.09572-.7659l-6.8936-.43086q-.14358.86167-.19144,1.508-.04719.64631-.04787,1.31652a11.64641,11.64641,0,0,0,.40683,3.25525,6.03505,6.03505,0,0,0,1.31652,2.39363,7.50638,7.50638,0,0,0,3.68621,2.13028,23.96494,23.96494,0,0,0,6.17546.64624h2.6808a3.34757,3.34757,0,0,1,2.32183.62241,2.91063,2.91063,0,0,1,.69421,2.05849H336.80394l-1.149-3.25535h-6.07973v3.30321q0,.62279.04786,1.508a8.66077,8.66077,0,0,0,.14369,1.31652,4.58993,4.58993,0,0,0-1.62773-2.274,3.66341,3.66341,0,0,0-1.96276-.59838H293.1687v5.88828h2.77453l-1.29049,2.96808a40.10919,40.10919,0,0,1,4.71545,1.17283,22.963,22.963,0,0,1,4.093,1.74738q-1.57976.67091-3.03986,1.41224-1.46094.74271-2.80056,1.43608-1.34072.69456-2.58507,1.4601L297.573,446.23a38.55419,38.55419,0,0,1,6.79787,1.43618A26.00216,26.00216,0,0,1,309.9,450.13162a21.85625,21.85625,0,0,1,4.57176,3.66218h7.2766V429.90562h3.7577a3.883,3.883,0,0,1,2.51328.78983,5.318,5.318,0,0,1,1.55583,2.22611V453.7938h7.2766V429.90562h7.24169a6.44265,6.44265,0,0,1,2.21252,1.29249,4.9232,4.9232,0,0,1,.89088,1.127,7.52421,7.52421,0,0,0-2.32705,1.60181,4.38813,4.38813,0,0,0-1.26866,3.11166,10.2304,10.2304,0,0,0,.16762,1.867,8.55624,8.55624,0,0,0,.50266,1.72335l6.84563-.86162q-.09641-.23905-.19145-.55052a2.014,2.014,0,0,1-.09572-.59849,1.66211,1.66211,0,0,1,.35907-1.101,2.17621,2.17621,0,0,1,1.101-.64624,6.926,6.926,0,0,1,1.891-.21549,5.32377,5.32377,0,0,1,3.30311.88566,2.841,2.841,0,0,1,1.149,2.36969,3.02953,3.02953,0,0,1-.83779,2.20208,5.41673,5.41673,0,0,1-2.51328,1.31652,16.61244,16.61244,0,0,1-4.26066.45479,26.70077,26.70077,0,0,1-3.59039-.23931q-1.77162-.23906-3.4468-.62242l-1.1489,5.60112a53.1323,53.1323,0,0,1,6.48667,1.31641,58.09954,58.09954,0,0,1,5.96008,1.96276,57.22508,57.22508,0,0,1,5.45742,2.46542,53.433,53.433,0,0,1,4.83511,2.82449l3.68611-5.07442a34.32468,34.32468,0,0,0-4.5-2.657A39.36982,39.36982,0,0,0,360.14108,447.54645Zm-48.25435-2.53721a24.88008,24.88008,0,0,0-4.45211-2.15421,22.23011,22.23011,0,0,0-3.59457-.97929c.63517-.33786,1.274-.67384,1.96695-.98348q1.3396-.598,2.68079-1.17283,1.33962-.5745,2.53722-1.00531l-.47873-4.78725a25.016,25.016,0,0,0-2.20207-1.67548,22.17734,22.17734,0,0,0-2.53721-1.484,20.19118,20.19118,0,0,0-2.11826-.86173H314.4718v13.59565q0,1.53273.04787,2.6809c.01285.31392.08328.62785.11777.94177A17.587,17.587,0,0,0,311.88673,445.00924Z" transform="translate(0.33691 -0.18781)" />
+      </g>
+    </g>
+  )
+}
+
+export default Bn
