+++ src/pages/HomePage/Mask/Narrowscreen/eng/index.tsx
@@ -1,0 +1,26 @@
+const En = () => {
+  return (
+    <g>
+      <g id="mask">
+        <path className="masking__layer" d="M292.36426,430.53809a5.15754,5.15754,0,0,0-3.45508,1.1748,5.30312,5.30312,0,0,0-1.76172,3.19629h10.43359a5.3091,5.3091,0,0,0-1.76269-3.19629A5.16024,5.16024,0,0,0,292.36426,430.53809Z" />
+        <path className="masking__layer" d="M251.21582,486.93848a5.15754,5.15754,0,0,0-3.45508,1.1748,5.30312,5.30312,0,0,0-1.76172,3.19629h10.4336a5.30916,5.30916,0,0,0-1.7627-3.19629A5.16022,5.16022,0,0,0,251.21582,486.93848Z" />
+        <path className="masking__layer" d="M320.77441,486.93848a5.15749,5.15749,0,0,0-3.45507,1.1748,5.30312,5.30312,0,0,0-1.76172,3.19629h10.43359a5.3092,5.3092,0,0,0-1.76269-3.19629A5.16027,5.16027,0,0,0,320.77441,486.93848Z" />
+        <path className="masking__layer" d="M0,.18781V599.81219H500V.18781ZM219.56848,136.50427a97.97244,97.97244,0,0,1,16.4784-27.069,74.32659,74.32659,0,0,1,8.98291-9.023,7.90872,7.90872,0,0,1,10.13989,0,74.33236,74.33236,0,0,1,8.98291,9.023,97.971,97.971,0,0,1,16.47833,27.069,96.68225,96.68225,0,0,1,7.08984,36.092,93.94115,93.94115,0,0,1-6.83648,36.092,96.07084,96.07084,0,0,1-9.6369,18.04615,97.05608,97.05608,0,0,1-6.95868,9.023,73.69406,73.69406,0,0,1-9.119,9.02282,7.989,7.989,0,0,1-10.14,0,73.68537,73.68537,0,0,1-9.119-9.02294,97.04714,97.04714,0,0,1-6.95868-9.02289,96.07024,96.07024,0,0,1-9.63684-18.04615,93.93991,93.93991,0,0,1-6.83655-36.092A96.68178,96.68178,0,0,1,219.56848,136.50427ZM246.68066,416.9082a4.55957,4.55957,0,0,1-1.5039,3.501,6.35616,6.35616,0,0,1-7.89649.02344,4.55873,4.55873,0,0,1,0-6.76758,5.65553,5.65553,0,0,1,3.94825-1.36328,5.85084,5.85084,0,0,1,3.94824,1.293A4.17281,4.17281,0,0,1,246.68066,416.9082Zm-28.623,14.85157a4.77905,4.77905,0,0,0-3.68946,1.5039,6.455,6.455,0,0,0-1.38671,4.51172v12.31445h-8.92969V424.61621h8.50683v2.77246a10.09618,10.09618,0,0,1,3.57227-2.373,12.104,12.104,0,0,1,4.51172-.82325,10.47045,10.47045,0,0,1,7.63769,2.82032q2.89014,2.81982,2.89063,8.50683v14.57031h-8.93067V436.97656Q222.24023,431.76025,218.05762,431.75977ZM102.3833,331.87164a97.96978,97.96978,0,0,1-27.069-16.47827,74.32612,74.32612,0,0,1-9.02295-8.98291,7.90874,7.90874,0,0,1,0-10.14,74.32946,74.32946,0,0,1,9.02295-8.98279,97.973,97.973,0,0,1,27.069-16.47839,96.6816,96.6816,0,0,1,36.092-7.08985,93.9377,93.9377,0,0,1,36.092,6.83655,96.06838,96.06838,0,0,1,18.04615,9.63684,97.0474,97.0474,0,0,1,9.02288,6.95868,73.686,73.686,0,0,1,9.023,9.119,7.98911,7.98911,0,0,1,0,10.14,73.69524,73.69524,0,0,1-9.02288,9.119,97.05608,97.05608,0,0,1-9.023,6.95868,96.06841,96.06841,0,0,1-18.04615,9.6369,93.9412,93.9412,0,0,1-36.092,6.83649A96.6814,96.6814,0,0,1,102.3833,331.87164Zm84.04346,117.84223a11.04681,11.04681,0,0,1-4.18262.79883,10.68114,10.68114,0,0,1-7.87207-2.91407q-2.93847-2.91357-2.9375-8.74218V424.61621h8.92969v12.7832a6.5532,6.5532,0,0,0,1.08105,4.1836,3.83915,3.83915,0,0,0,3.14844,1.31543,4.44654,4.44654,0,0,0,3.45508-1.48047,6.59088,6.59088,0,0,0,1.33887-4.53516v-12.2666h8.93066v25.47363h-8.50684v-2.72558A9.15831,9.15831,0,0,1,186.42676,449.71387ZM227.669,506.49023h-8.64844L214.22656,492.625l-4.98144,13.86523h-8.64844l-9.07031-25.47363h8.46l5.2168,15.416,5.40527-15.416h7.61328l5.26465,15.55664,5.4043-15.55664h7.8496Zm9.09472-81.874h8.92969v25.47363h-8.92969Zm27.89356,71.39258H246.13965a5.1835,5.1835,0,0,0,2.18555,3.03223,7.29505,7.29505,0,0,0,4.06543,1.05761,8.78321,8.78321,0,0,0,3.17285-.51758,9.37256,9.37256,0,0,0,2.70215-1.6914l4.70019,4.88769q-3.71339,4.13673-10.85742,4.13575a17.33243,17.33243,0,0,1-7.84863-1.69141,12.86092,12.86092,0,0,1-.28223-22.93652,16.03211,16.03211,0,0,1,14.09961-.11719,11.9844,11.9844,0,0,1,4.91211,4.55859,13.38565,13.38565,0,0,1,1.80957,7.07324Q264.79883,493.895,264.65723,496.00879Zm-6.08692-45.91895-10.48047-25.47363h9.165l6.10938,15.55664,6.43945-15.55664h8.50684l-10.52832,25.47363Zm44.43848,52.75782A9.4872,9.4872,0,0,1,298.709,505.832a19.28019,19.28019,0,0,1-6.83886,1.08106,27.47253,27.47253,0,0,1-6.29785-.72852,16.62536,16.62536,0,0,1-4.98243-1.90332l2.67969-6.11035a15.08508,15.08508,0,0,0,4.13574,1.71582,18.49356,18.49356,0,0,0,4.74707.63476,7.6,7.6,0,0,0,3.07813-.44628,1.40088,1.40088,0,0,0,.917-1.293q0-.84669-1.10449-1.19825a22.90611,22.90611,0,0,0-3.502-.68164,34.49922,34.49922,0,0,1-5.19336-.9873,7.99812,7.99812,0,0,1-3.68946-2.25586,6.44925,6.44925,0,0,1-1.55078-4.60547,6.96,6.96,0,0,1,1.457-4.34766,9.66175,9.66175,0,0,1,4.25293-3.00878,18.24691,18.24691,0,0,1,6.69825-1.1045,27.072,27.072,0,0,1,5.498.56446,15.83038,15.83038,0,0,1,4.55957,1.59765l-2.67871,6.06348a14.85851,14.85851,0,0,0-7.332-1.87988,7.31039,7.31039,0,0,0-3.10254.49316,1.47119,1.47119,0,0,0-.9873,1.293q0,.8921,1.08105,1.24512a24.63884,24.63884,0,0,0,3.57227.72851,39.45368,39.45368,0,0,1,5.21679,1.08106,7.97263,7.97263,0,0,1,3.59571,2.23242,6.35819,6.35819,0,0,1,1.52734,4.53516A6.89008,6.89008,0,0,1,303.00879,502.84766Zm2.79687-63.23926H287.28809a5.18352,5.18352,0,0,0,2.18554,3.03223,7.2951,7.2951,0,0,0,4.06543,1.05761,8.78313,8.78313,0,0,0,3.17285-.51758,9.37243,9.37243,0,0,0,2.70215-1.6914l4.7002,4.88769q-3.71339,4.13673-10.85742,4.13575a17.33244,17.33244,0,0,1-7.84864-1.69141,12.86093,12.86093,0,0,1-.28222-22.93652,16.03211,16.03211,0,0,1,14.09961-.11719,11.98446,11.98446,0,0,1,4.91211,4.55859,13.38565,13.38565,0,0,1,1.80957,7.07324Q305.94727,437.49463,305.80566,439.6084Zm28.41016,56.40039H315.69824a5.1835,5.1835,0,0,0,2.18555,3.03223,7.29507,7.29507,0,0,0,4.06543,1.05761,8.78313,8.78313,0,0,0,3.17285-.51758,9.37243,9.37243,0,0,0,2.70215-1.6914l4.70019,4.88769q-3.71337,4.13673-10.85742,4.13575a17.33243,17.33243,0,0,1-7.84863-1.69141,12.86092,12.86092,0,0,1-.28223-22.93652,16.03211,16.03211,0,0,1,14.09961-.11719,11.9844,11.9844,0,0,1,4.91211,4.55859,13.38555,13.38555,0,0,1,1.80957,7.07324Q334.35742,493.895,334.21582,496.00879ZM322.13867,425.0625a13.47022,13.47022,0,0,1,5.00488-.87012v8.03711a18.29985,18.29985,0,0,0-2.0205-.14062,6.22335,6.22335,0,0,0-4.48828,1.55078,6.28993,6.28993,0,0,0-1.62207,4.74707v11.70312H310.083V424.61621h8.50683v3.05469A8.31392,8.31392,0,0,1,322.13867,425.0625Zm21.14942,16.12109a22.906,22.906,0,0,0-3.502-.68164,34.49922,34.49922,0,0,1-5.19336-.9873,7.99809,7.99809,0,0,1-3.68945-2.25586,6.44925,6.44925,0,0,1-1.55078-4.60547,6.96,6.96,0,0,1,1.457-4.34766,9.66164,9.66164,0,0,1,4.25293-3.00878,18.24689,18.24689,0,0,1,6.69824-1.1045,27.072,27.072,0,0,1,5.498.56446,15.83038,15.83038,0,0,1,4.55957,1.59765l-2.67871,6.06348a14.85856,14.85856,0,0,0-7.332-1.87988,7.31039,7.31039,0,0,0-3.10254.49316,1.47121,1.47121,0,0,0-.98731,1.293q0,.89209,1.08106,1.24512a24.63894,24.63894,0,0,0,3.57226.72851,39.45387,39.45387,0,0,1,5.2168,1.08106,7.97254,7.97254,0,0,1,3.5957,2.23242,6.35814,6.35814,0,0,1,1.52735,4.53515,6.89009,6.89009,0,0,1-1.457,4.30079,9.4872,9.4872,0,0,1-4.29981,2.98437,19.28021,19.28021,0,0,1-6.83887,1.08106,27.47259,27.47259,0,0,1-6.29785-.72852,16.62528,16.62528,0,0,1-4.98242-1.90332l2.67969-6.11035a15.085,15.085,0,0,0,4.13574,1.71582,18.49356,18.49356,0,0,0,4.74707.63476,7.5998,7.5998,0,0,0,3.07813-.44629,1.40087,1.40087,0,0,0,.917-1.293Q344.39258,441.53515,343.28809,441.18359Zm21.10156,54.8252H345.87207a5.1835,5.1835,0,0,0,2.18555,3.03223,7.29505,7.29505,0,0,0,4.06543,1.05761,8.78317,8.78317,0,0,0,3.17285-.51758,9.37243,9.37243,0,0,0,2.70215-1.6914l4.70019,4.88769q-3.71337,4.13673-10.85742,4.13575a17.33243,17.33243,0,0,1-7.84863-1.69141,12.86092,12.86092,0,0,1-.28223-22.93652,16.03211,16.03211,0,0,1,14.09961-.11719,11.9844,11.9844,0,0,1,4.91211,4.55859,13.38565,13.38565,0,0,1,1.80957,7.07324Q364.53125,493.895,364.38965,496.00879Zm11.86816,9.40039a5.59924,5.59924,0,0,1-7.63672,0,5.076,5.076,0,0,1-1.55175-3.80664,5.02352,5.02352,0,0,1,1.52832-3.80762,5.79466,5.79466,0,0,1,7.68457,0,5.53819,5.53819,0,0,1-.02442,7.61426Zm6.2041-65.80078H363.94434a5.18352,5.18352,0,0,0,2.18554,3.03223,7.2951,7.2951,0,0,0,4.06543,1.05761,8.78313,8.78313,0,0,0,3.17285-.51758,9.37243,9.37243,0,0,0,2.70215-1.6914l4.7002,4.88769q-3.71339,4.13673-10.85742,4.13575a17.33244,17.33244,0,0,1-7.84864-1.69141,12.86093,12.86093,0,0,1-.28222-22.93652,16.03211,16.03211,0,0,1,14.09961-.11719,11.98446,11.98446,0,0,1,4.91211,4.55859,13.38565,13.38565,0,0,1,1.80957,7.07324Q382.60352,437.49463,382.46191,439.6084Zm51.44623-133.19794a74.3259,74.3259,0,0,1-9.023,8.98291,97.9689,97.9689,0,0,1-27.069,16.47827,96.68094,96.68094,0,0,1-36.092,7.08991,93.94068,93.94068,0,0,1-36.092-6.83649,96.06851,96.06851,0,0,1-18.0462-9.6369,97.05262,97.05262,0,0,1-9.02295-6.95868,73.69532,73.69532,0,0,1-9.02289-9.119,7.98911,7.98911,0,0,1,0-10.14,73.69043,73.69043,0,0,1,9.023-9.119,97.05417,97.05417,0,0,1,9.02289-6.95868,96.06849,96.06849,0,0,1,18.0462-9.63684,93.93718,93.93718,0,0,1,36.092-6.83655,96.68105,96.68105,0,0,1,36.092,7.08985,97.97217,97.97217,0,0,1,27.069,16.47839,74.32923,74.32923,0,0,1,9.023,8.98279A7.90888,7.90888,0,0,1,433.90814,306.41046Z" />
+        <path className="masking__layer" d="M369.02051,430.53809a5.15754,5.15754,0,0,0-3.45508,1.1748,5.30312,5.30312,0,0,0-1.76172,3.19629H374.2373a5.3091,5.3091,0,0,0-1.76269-3.19629A5.16024,5.16024,0,0,0,369.02051,430.53809Z" />
+        <path className="masking__layer" d="M350.94824,486.93848a5.15754,5.15754,0,0,0-3.45508,1.1748,5.30316,5.30316,0,0,0-1.76171,3.19629H356.165a5.30916,5.30916,0,0,0-1.7627-3.19629A5.16022,5.16022,0,0,0,350.94824,486.93848Z" />
+      </g>
+      <g id="visible">
+        <path className="visible__layer" d="M74.61523,423.39355H64.08789v-6.2041H92.75781v6.2041H82.22949v26.69629H74.61523Z" />
+        <path className="visible__layer" d="M119.28906,427.24805q2.89014,2.81982,2.89063,8.36621v14.47558h-7.332V436.74219a6.59708,6.59708,0,0,0-1.31641-4.48926,4.82924,4.82924,0,0,0-3.80664-1.48047,5.83258,5.83258,0,0,0-4.418,1.71582,7.10094,7.10094,0,0,0-1.64551,5.09961v12.50195h-7.332v-34.874h7.332v12.21973a9.76332,9.76332,0,0,1,3.52539-2.23242,12.754,12.754,0,0,1,4.51172-.7754A10.40925,10.40925,0,0,1,119.28906,427.24805Z" />
+        <path className="visible__layer" d="M153.71582,439.51465H134.58691a5.9529,5.9529,0,0,0,2.44434,3.71289,8.08867,8.08867,0,0,0,4.79395,1.36328,9.68761,9.68761,0,0,0,3.501-.58789,8.44968,8.44968,0,0,0,2.84375-1.85645l3.90137,4.23047q-3.57276,4.08838-10.43457,4.08887a16.48769,16.48769,0,0,1-7.56641-1.66894,12.2245,12.2245,0,0,1-5.07617-4.62891,12.74911,12.74911,0,0,1-1.78613-6.7207,12.92711,12.92711,0,0,1,1.76269-6.69825,12.34413,12.34413,0,0,1,4.84082-4.65234,14.98727,14.98727,0,0,1,13.60645-.07129,11.68281,11.68281,0,0,1,4.72363,4.583,13.66818,13.66818,0,0,1,1.71582,6.93262Q153.85742,437.68165,153.71582,439.51465Zm-17.1543-8.13086a5.99912,5.99912,0,0,0-2.06836,3.85449h12.45508a6.06522,6.06522,0,0,0-2.06836-3.831,6.18412,6.18412,0,0,0-4.13574-1.4336A6.31017,6.31017,0,0,0,136.56152,431.38379Z" />
+        <path className="visible__layer" d="M400.93262,420.10352a3.98129,3.98129,0,0,1,0-5.82813,4.65585,4.65585,0,0,1,3.29-1.1748,4.76789,4.76789,0,0,1,3.29,1.12793,3.61882,3.61882,0,0,1,1.26953,2.82031,3.99222,3.99222,0,0,1-1.26953,3.03125,4.59845,4.59845,0,0,1-3.29,1.19922A4.64777,4.64777,0,0,1,400.93262,420.10352Zm-.376,4.70019h7.332v25.28613h-7.332Z" />
+        <path className="visible__layer" d="M417.05371,449.69043a16.07034,16.07034,0,0,1-4.79394-1.9502l2.44336-5.26464a15.34467,15.34467,0,0,0,4.08886,1.76269,17.11538,17.11538,0,0,0,4.7002.68164q4.65381,0,4.65332-2.30273a1.6314,1.6314,0,0,0-1.26953-1.55176,18.8799,18.8799,0,0,0-3.90039-.79883,37.343,37.343,0,0,1-5.12305-1.08105,7.93823,7.93823,0,0,1-3.50195-2.16211,6.14781,6.14781,0,0,1-1.48047-4.418,6.91824,6.91824,0,0,1,1.38672-4.25293,8.92347,8.92347,0,0,1,4.042-2.89063,17.27636,17.27636,0,0,1,6.27441-1.03418,24.63379,24.63379,0,0,1,5.335.5879,14.95086,14.95086,0,0,1,4.39453,1.62109l-2.44433,5.2168a14.59813,14.59813,0,0,0-7.28516-1.87989,7.41348,7.41348,0,0,0-3.52539.65821,1.93569,1.93569,0,0,0-1.1748,1.69238,1.67842,1.67842,0,0,0,1.26953,1.64453,22.43543,22.43543,0,0,0,4.042.89355,43.344,43.344,0,0,1,5.07617,1.1045,7.63417,7.63417,0,0,1,3.43067,2.13769,6.072,6.072,0,0,1,1.457,4.32422,6.67632,6.67632,0,0,1-1.41016,4.18359,8.99727,8.99727,0,0,1-4.1123,2.84278,18.32435,18.32435,0,0,1-6.416,1.01074A24.57138,24.57138,0,0,1,417.05371,449.69043Z" />
+        <path className="visible__layer" d="M143.68164,483.625q3.14795,2.79639,3.14844,8.43652v14.42871h-6.86231v-3.14941q-2.06835,3.52589-7.708,3.52539a11.95445,11.95445,0,0,1-5.05176-.9873,7.52275,7.52275,0,0,1-3.2666-2.72559,7.09339,7.09339,0,0,1-1.12793-3.94824,6.58833,6.58833,0,0,1,2.65527-5.5459q2.65429-2.02,8.20117-2.02051h5.82813a4.67328,4.67328,0,0,0-1.457-3.68945,6.41542,6.41542,0,0,0-4.3711-1.293,12.76874,12.76874,0,0,0-3.97168.63477,10.53373,10.53373,0,0,0-3.31347,1.71484l-2.63184-5.123a15.91521,15.91521,0,0,1,4.959-2.25586,22.225,22.225,0,0,1,5.94532-.79882Q140.53125,480.82813,143.68164,483.625Zm-6.251,17.43652A4.7574,4.7574,0,0,0,139.49805,498.5v-2.585h-5.0293q-4.51172,0-4.51172,2.96094a2.63556,2.63556,0,0,0,1.10449,2.23242,4.9701,4.9701,0,0,0,3.03125.82324A6.387,6.387,0,0,0,137.43066,501.06152Z" />
+        <path className="visible__layer" d="M153.45605,471.61621h7.332v34.874h-7.332Z" />
+        <path className="visible__layer" d="M167.60352,471.61621h7.332v34.874h-7.332Z" />
+      </g>
+    </g>
+  )
+}
+
+export default En
