+++ src/pages/HomePage/Mask/Widescreen/bn/index.tsx
@@ -1,0 +1,31 @@
+const Bn = () => {
+  return (
+    <g>
+      <g id="mask">
+        <path className="masking__layer" d="M-.67383-138.11322V558.002H900V-138.11322ZM173.20929,376.39691a20.39656,20.39656,0,0,1-3.91.35163,15.988,15.988,0,0,1-7.56311-1.69977,11.60407,11.60407,0,0,1-4.99457-5.18762,18.98214,18.98214,0,0,1-1.76953-8.64582,20.95066,20.95066,0,0,1,1.1131-6.82879,23.17857,23.17857,0,0,1,3.05389-5.97883,25.71738,25.71738,0,0,1,4.50928-4.86517,26.07138,26.07138,0,0,1,5.5368-3.54626h7.36334v7.20977h-3.48193a24.4334,24.4334,0,0,0-3.68164,2.6377,20.20967,20.20967,0,0,0-2.96814,3.19452,14.15212,14.15212,0,0,0-1.9978,3.75159,12.66241,12.66241,0,0,0-.71357,4.24963,10.80941,10.80941,0,0,0,.9704,5.041,5.745,5.745,0,0,0,2.42591,2.54981,6.36346,6.36346,0,0,0,2.88256.73267,11.0451,11.0451,0,0,0,2.02625-.17585,7.02692,7.02692,0,0,0,2.02643-.762l2.797,6.68225A13.6898,13.6898,0,0,1,173.20929,376.39691Zm132.55066.0293a4.32432,4.32432,0,0,1-3.1966,1.26019,4.31154,4.31154,0,0,1-4.50921-4.51337,4.33677,4.33677,0,0,1,1.25574-3.19464,4.39321,4.39321,0,0,1,3.25347-1.26025,4.30424,4.30424,0,0,1,4.45234,4.45489A4.41281,4.41281,0,0,1,305.76,376.42621Zm23.80237-29.22016H324.368v29.24946h-8.67615a25.26166,25.26166,0,0,0-4.96606-5.15833,28.50205,28.50205,0,0,0-6.19312-3.634,34.95515,34.95515,0,0,0-7.56317-2.22748l-3.19641-7.97174a117.84164,117.84164,0,0,1,11.35895-5.36341,92.60758,92.60758,0,0,1,10.90222-3.6853v-1.20917h-26.314a25.74684,25.74684,0,0,0-3.71027,2.6377,19.566,19.566,0,0,0-2.9967,3.19452,14.15764,14.15764,0,0,0-1.99781,3.75159,12.66557,12.66557,0,0,0-.71344,4.24963,10.81105,10.81105,0,0,0,.97028,5.041,5.74667,5.74667,0,0,0,2.4259,2.54981,6.36353,6.36353,0,0,0,2.88257.73267,11.04765,11.04765,0,0,0,2.02643-.17585,7.02584,7.02584,0,0,0,2.02624-.762l2.797,6.68225a13.68905,13.68905,0,0,1-3.62457,1.28955,20.39731,20.39731,0,0,1-3.91.35163,16.03453,16.03453,0,0,1-7.53466-1.69977,11.56585,11.56585,0,0,1-5.023-5.18762,18.98044,18.98044,0,0,1-1.76953-8.64582,18.20033,18.20033,0,0,1,5.2514-12.92486,17.53552,17.53552,0,0,1,1.32452-1.21326c-.33679.028-.69128.0517-1.06768.07025q-1.1706.059-3.05378.05859H248.05182v.45722a12.07542,12.07542,0,0,0,1.855,2.0047,9.4635,9.4635,0,0,0,3.11084,1.69989,12.46509,12.46509,0,0,0,3.9386.58612h1.36987a15.74407,15.74407,0,0,1,4.39527.49817,6.41769,6.41769,0,0,1,2.79693,1.67059,5.1455,5.1455,0,0,1,1.51257,2.286,9.30694,9.30694,0,0,1,.48523,3.048,13.7476,13.7476,0,0,1-1.3128,6.09607,17.96809,17.96809,0,0,1-3.48193,4.86517,18.71811,18.71811,0,0,1-5.394,3.57556,16.27107,16.27107,0,0,1-6.64984,1.34815,15.76043,15.76043,0,0,1-4.7663-.61542,8.19149,8.19149,0,0,1-5.53674-5.09955,17.17534,17.17534,0,0,1-.8847-6.35992v-16.0608h-8.96094v29.24946h-8.67621V350.89886a6.51271,6.51271,0,0,0-1.8551-2.7257,4.54874,4.54874,0,0,0-2.99671-.96711h-4.33764v29.24946h-8.67627v-22.134c-.56317-.68365-1.1167-1.36963-1.7124-2.04522a26.97765,26.97765,0,0,0-2.91114-2.8429,13.85561,13.85561,0,0,0-3.25348-2.05151,9.03273,9.03273,0,0,0-3.71032-.762,7.06581,7.06581,0,0,0-4.25244,1.11371,22.91223,22.91223,0,0,0-1.7862,1.48377c.30707-.12952.61182-.26331.93006-.37018a11.20075,11.20075,0,0,1,3.51043-.46881,8.5764,8.5764,0,0,1,4.16687.93774,6.71156,6.71156,0,0,1,2.65418,2.57923,7.64184,7.64184,0,0,1,.94183,3.86865,10.22625,10.22625,0,0,1-2.56866,6.62359,25.6957,25.6957,0,0,1-7.64875,5.86152l-4.79473-6.858a30.62643,30.62643,0,0,0,5.0802-2.90149,3.46393,3.46393,0,0,0,1.65533-2.667,1.73416,1.73416,0,0,0-.51373-1.34814,2.41519,2.41519,0,0,0-1.65534-.469,7.26361,7.26361,0,0,0-2.19763.35175,13.75338,13.75338,0,0,0-2.483,1.11365l-6.50714-5.45136a41.7329,41.7329,0,0,1,4.36658-5.36322,19.39,19.39,0,0,1,4.88043-3.69293,12.5878,12.5878,0,0,1,5.93634-1.37744,13.02136,13.02136,0,0,1,4.93744.93786,18.78587,18.78587,0,0,1,4.62347,2.78424q1.17582.93906,2.43317,2.13031c-.01148-.23681-.01984-.44555-.03577-.72345q-.08561-1.49478-.08563-3.31188v-5.39271h7.30621l1.31287,3.986h5.25134v.00885a4.29341,4.29341,0,0,1,2.2829.72381,5.61224,5.61224,0,0,1,1.94067,2.7843,10.88974,10.88974,0,0,1-.1712-1.61193q-.05758-1.0839-.05707-1.84644v-4.04456h7.24915l1.37,3.986h99.09052Zm81.0971-100.79559a73.69524,73.69524,0,0,1-9.02288,9.119,97.05608,97.05608,0,0,1-9.023,6.95868,96.06841,96.06841,0,0,1-18.04615,9.6369,93.9412,93.9412,0,0,1-36.092,6.83649,96.6814,96.6814,0,0,1-36.092-7.08991,97.96978,97.96978,0,0,1-27.069-16.47827,74.32612,74.32612,0,0,1-9.023-8.98291,7.90874,7.90874,0,0,1,0-10.14,74.32946,74.32946,0,0,1,9.023-8.98279,97.973,97.973,0,0,1,27.069-16.47839,96.6816,96.6816,0,0,1,36.092-7.08985,93.9377,93.9377,0,0,1,36.092,6.83655,96.06838,96.06838,0,0,1,18.04615,9.63684,97.0474,97.0474,0,0,1,9.02288,6.95868,73.686,73.686,0,0,1,9.023,9.119A7.98911,7.98911,0,0,1,410.65942,246.41046Zm25.25129-70.65307a97.04714,97.04714,0,0,1-6.95868-9.02289,96.07024,96.07024,0,0,1-9.63684-18.04615,93.93991,93.93991,0,0,1-6.83655-36.092,96.68178,96.68178,0,0,1,7.08984-36.092,97.97244,97.97244,0,0,1,16.4784-27.069,74.32659,74.32659,0,0,1,8.98291-9.023,7.90872,7.90872,0,0,1,10.13989,0,74.33236,74.33236,0,0,1,8.98291,9.023,97.971,97.971,0,0,1,16.47833,27.069,96.68225,96.68225,0,0,1,7.08984,36.092,93.94115,93.94115,0,0,1-6.83648,36.092,96.07084,96.07084,0,0,1-9.6369,18.04615,97.05608,97.05608,0,0,1-6.95868,9.023,73.69406,73.69406,0,0,1-9.119,9.02282,7.989,7.989,0,0,1-10.14,0A73.68537,73.68537,0,0,1,435.91071,175.75739ZM618.87292,347.20605H599.12311v8.49933a3.0391,3.0391,0,0,0,.314,1.49469,1.2224,1.2224,0,0,0,1.17017.55683,2.91876,2.91876,0,0,0,1.5412-.49823,6.838,6.838,0,0,0,1.56964-1.40674,24.946,24.946,0,0,0,1.56976-2.08081l7.02088-1.28967a23.13434,23.13434,0,0,1,3.08227,5.5979,16.71534,16.71534,0,0,1,1.02753,5.83227,11.81268,11.81268,0,0,1-1.71247,6.15466,12.22637,12.22637,0,0,1-5.2514,4.63068,20.54976,20.54976,0,0,1-9.13287,1.75855,26.517,26.517,0,0,1-8.41931-1.31885,21.334,21.334,0,0,1-7.30622-4.2204,27.32392,27.32392,0,0,1-5.96484-7.76666l6.9638-3.92719a31.01345,31.01345,0,0,0,3.90992,5.33393,14.077,14.077,0,0,0,4.48083,3.28259,13.24814,13.24814,0,0,0,5.53675,1.11377,10.35133,10.35133,0,0,0,6.136-1.5827,5.28834,5.28834,0,0,0,2.19763-4.57215,7.03317,7.03317,0,0,0-.14215-1.18677,15.65668,15.65668,0,0,1-4.13885,2.44715,12.85008,12.85008,0,0,1-4.62348.84991,10.80749,10.80749,0,0,1-3.68164-.58618,6.45406,6.45406,0,0,1-2.71136-1.81714,6.859,6.859,0,0,1-.48309-.69873,54.27236,54.27236,0,0,0-6.19526-4.04919,13.59611,13.59611,0,0,0-6.50714-1.87567,2.8707,2.8707,0,0,0-1.8551.49823,1.51334,1.51334,0,0,0-.59936,1.20166,1.7894,1.7894,0,0,0,.45666,1.26025,4.52212,4.52212,0,0,0,1.59821.96722l-3.71021,6.68219a14.05794,14.05794,0,0,1-5.27991-3.898,8.241,8.241,0,0,1-1.91217-5.30481,7.81662,7.81662,0,0,1,1.42694-4.86511,8.06791,8.06791,0,0,1,3.7959-2.72558,14.85308,14.85308,0,0,1,5.05158-.85,18.83234,18.83234,0,0,1,4.90893.6449,21.05049,21.05049,0,0,1,4.76624,1.96362,23.78265,23.78265,0,0,1,2.54352,1.66138c-.00977-.21753-.02509-.4018-.03205-.63568q-.02957-.93687-.02856-1.75837v-3.517H561.9646v29.24946h-8.67621V350.89886a6.51271,6.51271,0,0,0-1.8551-2.7257,4.54874,4.54874,0,0,0-2.99671-.96711H530.24005V352.035a35.51218,35.51218,0,0,0,4.05274-.69647,48.78064,48.78064,0,0,0,5.30847-1.61194l6.2218,5.15814a18.9297,18.9297,0,0,0-1.6839,4.89454,25.69854,25.69854,0,0,0-.5423,5.1289,30.82585,30.82585,0,0,0,.34241,4.57208,30.11012,30.11012,0,0,0,1.08459,4.63062l-8.562,1.69989a37.96325,37.96325,0,0,1-1.1701-5.48059,40.05127,40.05127,0,0,1-.3711-5.422,27.08543,27.08543,0,0,1,.62787-5.92029c.04486-.20117.11853-.38336.16822-.58129a28.42315,28.42315,0,0,1-4.19239.78656c-.40857.04248-.85687.06317-1.2843.09545v17.16694h-8.619v-.453a40.359,40.359,0,0,0-2.91113-3.826,15.07252,15.07252,0,0,0-2.99664-2.69629,5.59194,5.59194,0,0,0-3.111-.99652,2.82379,2.82379,0,0,0-1.22723.23444,1.6836,1.6836,0,0,0-.742.67407,2.248,2.248,0,0,0-.25684,1.143,5.674,5.674,0,0,0,.1427,1.231,10.32345,10.32345,0,0,0,.4281,1.40686l-8.44794,1.28956q-.45757-1.28871-.77056-2.49122a9.79837,9.79837,0,0,1-.314-2.49115,7.36059,7.36059,0,0,1,.88477-3.75152,7.25727,7.25727,0,0,1,2.31176-2.46185,10.01584,10.01584,0,0,1,1.02576-.55408c.00012-.032.00171-.05847.00171-.0907a9.12812,9.12812,0,0,0-.314-2.22736,5.70838,5.70838,0,0,0-1.02746-2.13952,2.30829,2.30829,0,0,0-1.91217-.90863,1.78721,1.78721,0,0,0-1.34137.49829,1.97547,1.97547,0,0,0-.48517,1.43616,6.069,6.069,0,0,0,.11414,1.1723,4.22918,4.22918,0,0,0,.34253,1.055l-8.448,1.28955a13.99267,13.99267,0,0,1-.71344-2.25659,10.26975,10.26975,0,0,1-.25684-2.25678,7.23657,7.23657,0,0,1,.88471-3.69281,7.16248,7.16248,0,0,1,2.34027-2.46185,10.17558,10.17558,0,0,1,.95947-.50622,6.87013,6.87013,0,0,0-1.10211-3.53833,6.3193,6.3193,0,0,0-.75788-.87922H485.295v36.10749h-8.56213a5.62646,5.62646,0,0,0-2.05481-2.13941,7.0697,7.0697,0,0,0-3.539-.73266,19.54493,19.54493,0,0,0-2.79694.20508q-1.4274.20462-2.9967.43963a23.56113,23.56113,0,0,1-3.45337.23444,18.85575,18.85575,0,0,1-3.68164-.381,10.04209,10.04209,0,0,1-3.56751-1.46534,7.86841,7.86841,0,0,1-2.68273-3.10669,11.58842,11.58842,0,0,1-1.02747-5.24615q0-1.05506.08563-2.02222a11.27271,11.27271,0,0,1,.31384-1.84637l8.27668.93774q-.115.52763-.19977,1.08441a7.52624,7.52624,0,0,0-.08557,1.14307,3.22021,3.22021,0,0,0,1.19867,2.90149,5.38105,5.38105,0,0,0,3.0252.79132,14.50581,14.50581,0,0,0,2.36884-.17584q1.05561-.17578,2.112-.35175a13.73215,13.73215,0,0,1,2.2547-.17578,6.92364,6.92364,0,0,1,3.45325.82056,16.452,16.452,0,0,1,2.93963,2.16888v-1.17236a43.8473,43.8473,0,0,0-5.394-4.45477,31.098,31.098,0,0,0-5.93634-3.19471,39.84276,39.84276,0,0,0-7.44891-2.1394L454.7,357.46393a117.85516,117.85516,0,0,1,11.35895-5.36341,91.34707,91.34707,0,0,1,10.90234-3.67041v-1.22406H448.87781v-7.20977H549.23566a4.287,4.287,0,0,1,2.34033.73266,5.61224,5.61224,0,0,1,1.94067,2.7843,10.88974,10.88974,0,0,1-.1712-1.61193q-.05759-1.0839-.05707-1.84644v-4.04456h7.24915l1.37,3.986h56.96539Zm15.03522-100.79559a74.3259,74.3259,0,0,1-9.023,8.98291,97.9689,97.9689,0,0,1-27.069,16.47827,96.68094,96.68094,0,0,1-36.092,7.08991,93.94068,93.94068,0,0,1-36.092-6.83649,96.06851,96.06851,0,0,1-18.0462-9.6369,97.05262,97.05262,0,0,1-9.02295-6.95868,73.69532,73.69532,0,0,1-9.02289-9.119,7.98911,7.98911,0,0,1,0-10.14,73.69043,73.69043,0,0,1,9.023-9.119,97.05417,97.05417,0,0,1,9.02289-6.95868,96.06849,96.06849,0,0,1,18.0462-9.63684,93.93718,93.93718,0,0,1,36.092-6.83655,96.68105,96.68105,0,0,1,36.092,7.08985,97.97217,97.97217,0,0,1,27.069,16.47839,74.32923,74.32923,0,0,1,9.023,8.98279A7.90888,7.90888,0,0,1,633.90814,246.41046Zm111.29529,132.0437h-8.73334V341.99493h8.73334Z" />
+
+        <path className="masking__layer" d="M315.69189,363.56v-7.50464c-1.81842.54175-3.57324,1.12842-5.22283,1.78949q-2.99681,1.20208-6.25037,2.7843a19.0855,19.0855,0,0,1,4.538,1.52393,28.48341,28.48341,0,0,1,4.62348,2.75506,20.00408,20.00408,0,0,1,2.4566,2.14722c-.01849-.27118-.0423-.52118-.0592-.79907Q315.69192,364.84958,315.69189,363.56Z" />
+
+        <path className="masking__layer" d="M476.67584,362.56348v-6.52564c-1.83935.54639-3.61285,1.139-5.2799,1.807q-2.99662,1.20208-6.25025,2.7843a20.78992,20.78992,0,0,1,4.50928,1.49463,31.57237,31.57237,0,0,1,4.3667,2.46191,29.142,29.142,0,0,1,2.76556,2.149c-.02014-.35571-.04-.71246-.05432-1.09387Q476.6756,364.14621,476.67584,362.56348Z" />
+
+        <path className="masking__layer" d="M248.166,357.81555v5.92023a10.70278,10.70278,0,0,0,.19983,2.46185,2.3086,2.3086,0,0,0,.65644,1.17237,2.77762,2.77762,0,0,0,2.16894.762,6.39307,6.39307,0,0,0,2.73993-.58618,11.49219,11.49219,0,0,0,2.51148-1.64112,8.586,8.586,0,0,0,1.8551-2.22747,4.76576,4.76576,0,0,0,.65649-2.34467,1.86569,1.86569,0,0,0-.5708-1.52393,2.65633,2.65633,0,0,0-1.7124-.469h-.45667a10.92822,10.92822,0,0,1-2.9967-.381,13.24978,13.24978,0,0,1-4.909-2.8429c-.08807-.08227-.15088-.172-.23632-.25506.0122.15893.02429.29126.03662.46026Q248.16509,357.11225,248.166,357.81555Z" />
+
+        <path className="masking__layer" d="M521.621,362.91522V359.348c-1.27069-.0683-2.44336-.168-3.51043-.30157a30.92187,30.92187,0,0,1-5.12079-1.053c.06317.15979.15735.30194.21179.46686a12.13406,12.13406,0,0,1,.55176,3.23462,11.13673,11.13673,0,0,1,2.18817.66345,15.26163,15.26163,0,0,1,4.08124,2.6084,23.40108,23.40108,0,0,1,1.7008,1.68182c-.03473-.49683-.06219-.96912-.074-1.38867Q521.62005,364.2637,521.621,362.91522Z" />
+
+        <path className="masking__layer" d="M521.621,352.39581v-5.18976H503.1637c.07281.15687.15588.30518.22021.469a11.92792,11.92792,0,0,1,.67755,2.799c.255.03619.47089.11438.71448.1629a72.98124,72.98124,0,0,0,8.53992,1.34632C515.93652,352.24036,518.72162,352.3642,521.621,352.39581Z" />
+      </g>
+      <g id="visible">
+        <path className="visible__layer"
+          d="M423.71037,368.92529c-.1261-.044-.25481-.07806-.38129-.12118a14.0991,14.0991,0,0,0,4.49107-2.3993q2.82543-2.37382,2.82537-7.06316a9.88885,9.88885,0,0,0-1.42695-5.07037,10.44584,10.44584,0,0,0-4.595-3.98583,19.27927,19.27927,0,0,0-8.4193-1.55338c-.11613,0-.21257.021-.32733.02239a6.54443,6.54443,0,0,0-.61454-.87227,7.05067,7.05067,0,0,0-.55473-.55687H434.47v-7.20979H428.9903a17.95172,17.95172,0,0,0-.77067-4.98235,7.90747,7.90747,0,0,0-1.91216-3.16537,8.29594,8.29594,0,0,0-3.45337-1.78767,22.69336,22.69336,0,0,0-5.337-.49826h-3.25351a14.81745,14.81745,0,0,1-3.13949-.23454,3.83214,3.83214,0,0,1-1.59816-.762,3.4417,3.4417,0,0,1-.82774-1.43605,7.03152,7.03152,0,0,1-.25681-2.02234,5.17374,5.17374,0,0,1,.05707-.70337c.03738-.27293.07576-.58617.11414-.93779l-8.21956-.52756q-.17121,1.05505-.22828,1.8464-.05625.79134-.05706,1.612a14.62158,14.62158,0,0,0,.48508,3.98582,7.44843,7.44843,0,0,0,1.56974,2.93083,8.90241,8.90241,0,0,0,4.39525,2.60838,27.859,27.859,0,0,0,7.36329.79128h3.19644a3.91755,3.91755,0,0,1,2.76843.7621,3.6226,3.6226,0,0,1,.82774,2.52048H395.8844l-1.37-3.986h-7.24915v4.04456q0,.76256.05707,1.8464a10.87713,10.87713,0,0,0,.17133,1.612,5.61264,5.61264,0,0,0-1.94082-2.78432,4.28654,4.28654,0,0,0-2.34029-.73267H343.85605v7.20979h3.30821l-1.53872,3.6342a46.79913,46.79913,0,0,1,5.62246,1.43605,27.003,27.003,0,0,1,4.88032,2.13955q-1.88362.82147-3.62457,1.7292-1.742.90937-3.33924,1.75837-1.5986.85045-3.0823,1.78779l3.02523,7.50295a44.92884,44.92884,0,0,1,8.10543,1.75849,30.60717,30.60717,0,0,1,6.59274,3.01874,26.193,26.193,0,0,1,5.45113,4.48408H377.933V347.32532h4.48048a4.54881,4.54881,0,0,1,2.9967.96709,6.51285,6.51285,0,0,1,1.85509,2.72572v25.55661h8.67623V347.32532h8.63461a7.64069,7.64069,0,0,1,2.6381,1.58256,6.00661,6.00661,0,0,1,1.06224,1.37988,8.97054,8.97054,0,0,0-2.77466,1.9613,5.44915,5.44915,0,0,0-1.51268,3.81,12.85377,12.85377,0,0,0,.19986,2.286,10.69206,10.69206,0,0,0,.59934,2.11012l8.16237-1.055q-.11495-.29271-.22827-.67407a2.52656,2.52656,0,0,1-.11414-.73281,2.07109,2.07109,0,0,1,.42814-1.34814,2.58288,2.58288,0,0,1,1.31282-.79128,8.05316,8.05316,0,0,1,2.25469-.26384,6.226,6.226,0,0,1,3.93845,1.08442,3.51271,3.51271,0,0,1,1.37,2.90153,3.76532,3.76532,0,0,1-.99894,2.69628,6.4072,6.4072,0,0,1-2.99671,1.612,19.312,19.312,0,0,1-5.08019.55686,31.015,31.015,0,0,1-4.281-.293q-2.11239-.29271-4.10978-.76211l-1.36989,6.85818a61.8946,61.8946,0,0,1,7.73436,1.61186,67.81294,67.81294,0,0,1,13.61363,5.422,63.34374,63.34374,0,0,1,5.76513,3.45839l4.39512-6.21327a40.70087,40.70087,0,0,0-5.36553-3.25328A46.26542,46.26542,0,0,0,423.71037,368.92529Zm-57.53593-3.10664A29.341,29.341,0,0,0,360.866,363.181a25.96042,25.96042,0,0,0-4.286-1.19908c.75734-.41368,1.519-.82506,2.34528-1.20419q1.59727-.73222,3.19644-1.436,1.59727-.70342,3.02523-1.23093l-.5708-5.86166a29.92561,29.92561,0,0,0-2.62564-2.05151,26.29052,26.29052,0,0,0-3.02523-1.8171,23.71506,23.71506,0,0,0-2.52571-1.05513h12.85716v16.64693q0,1.87673.05707,3.28257c.01533.38438.09931.76876.14043,1.15314A21.04541,21.04541,0,0,0,366.17444,365.81865Z" />
+
+        <path className="visible__layer"
+          d="M717.42715,368.92529c-.12609-.044-.25481-.07806-.38128-.12118a14.097,14.097,0,0,0,4.49094-2.3993q2.82563-2.37382,2.8255-7.06316a9.88886,9.88886,0,0,0-1.427-5.07037,10.44581,10.44581,0,0,0-4.595-3.98583,19.27966,19.27966,0,0,0-8.4193-1.55338c-.11613,0-.21257.021-.32733.02239a6.54534,6.54534,0,0,0-.61454-.87227,7.04913,7.04913,0,0,0-.55474-.55687h19.76229v-7.20979H722.707a17.95181,17.95181,0,0,0-.77055-4.98235,7.90725,7.90725,0,0,0-1.91216-3.16537,8.29588,8.29588,0,0,0-3.45337-1.78767,22.69332,22.69332,0,0,0-5.337-.49826h-3.25351a14.81756,14.81756,0,0,1-3.1395-.23454,3.83214,3.83214,0,0,1-1.59816-.762,3.44309,3.44309,0,0,1-.82774-1.43605,7.03154,7.03154,0,0,1-.2568-2.02234,5.17374,5.17374,0,0,1,.05707-.70337c.0375-.27293.07575-.58617.11413-.93779l-8.21956-.52756q-.1712,1.05505-.22827,1.8464-.05627.79134-.05707,1.612a14.6166,14.6166,0,0,0,.4852,3.98582,7.44671,7.44671,0,0,0,1.56963,2.93083,8.90291,8.90291,0,0,0,4.39524,2.60838,27.859,27.859,0,0,0,7.36329.79128h3.19644a3.91751,3.91751,0,0,1,2.76843.7621,3.6226,3.6226,0,0,1,.82774,2.52048H689.60119l-1.36989-3.986H680.982v4.04456q0,.76256.0572,1.8464a10.86585,10.86585,0,0,0,.1712,1.612,5.61328,5.61328,0,0,0-1.94069-2.78432,4.28927,4.28927,0,0,0-2.3226-.73v-.00268H633.16662v7.20979h42.96361a4.5486,4.5486,0,0,1,2.99671.96709,6.51174,6.51174,0,0,1,1.85508,2.72572v25.55661h8.67624V347.32532h8.63461a7.64054,7.64054,0,0,1,2.63809,1.58256,6.00992,6.00992,0,0,1,1.06224,1.38,8.97118,8.97118,0,0,0-2.77466,1.96118,5.4492,5.4492,0,0,0-1.51268,3.81,12.85283,12.85283,0,0,0,.19987,2.286,10.6907,10.6907,0,0,0,.59934,2.11012l8.16237-1.055q-.11495-.29271-.22828-.67407a2.52688,2.52688,0,0,1-.11413-.73281,2.07114,2.07114,0,0,1,.42813-1.34814,2.58293,2.58293,0,0,1,1.31282-.79128,8.05321,8.05321,0,0,1,2.25469-.26384,6.22577,6.22577,0,0,1,3.93845,1.08442,3.5126,3.5126,0,0,1,1.36989,2.90153,3.76575,3.76575,0,0,1-.99882,2.69628,6.40712,6.40712,0,0,1-2.9967,1.612,19.31214,19.31214,0,0,1-5.08019.55686,31.01526,31.01526,0,0,1-4.281-.293q-2.11239-.29271-4.10978-.76211l-1.36988,6.85818a61.8946,61.8946,0,0,1,7.73436,1.61186,67.81253,67.81253,0,0,1,13.61362,5.422,63.32223,63.32223,0,0,1,5.765,3.45839l4.39524-6.21327a40.69956,40.69956,0,0,0-5.36553-3.25328A46.264,46.264,0,0,0,717.42715,368.92529Z" />
+        <path className="visible__layer"
+          d="M670.25827,350.813a15.46354,15.46354,0,0,0-7.50608-1.6706,16.022,16.022,0,0,0-5.16579.82058,8.39139,8.39139,0,0,0-3.90992,2.63769,7.45662,7.45662,0,0,0-1.48415,4.86513,11.33934,11.33934,0,0,0,.22827,2.08095,12.55513,12.55513,0,0,0,.79921,2.54977l8.16249-1.11373a9.38132,9.38132,0,0,1-.28546-1.11372,6.37649,6.37649,0,0,1-.11414-1.23094,2.35052,2.35052,0,0,1,.742-1.8757,2.76788,2.76788,0,0,1,1.88362-.64477,4.01844,4.01844,0,0,1,2.112.58616,3.92281,3.92281,0,0,1,1.51268,1.7292,6.46866,6.46866,0,0,1,.54215,2.78432,6.65911,6.65911,0,0,1-.51361,2.69628,4.44894,4.44894,0,0,1-1.56975,1.905,7.39764,7.39764,0,0,1-2.56857,1.08442,15.61875,15.61875,0,0,1-3.51044.35175,11.71117,11.71117,0,0,1-6.4786-1.8171,15.19992,15.19992,0,0,1-4.93752-5.68584,46.43878,46.43878,0,0,1-3.76725-9.96469l-8.04836,2.28606q2.39743,8.61648,6.02207,13.59889a20.89163,20.89163,0,0,0,8.27663,7.06329,24.38047,24.38047,0,0,0,10.07465,2.08081,20.30034,20.30034,0,0,0,8.9045-1.69989,11.61726,11.61726,0,0,0,5.16567-4.63059,12.83,12.83,0,0,0,1.68388-6.50643,13.40363,13.40363,0,0,0-1.59816-6.62363A11.50485,11.50485,0,0,0,670.25827,350.813Z" />
+
+      </g>
+    </g>
+  )
+}
+
+export default Bn
