+++ src/pages/HomePage/Mask/Widescreen/eng/index.tsx
@@ -1,0 +1,32 @@
+import React from 'react'
+
+const En = () => {
+  return (
+    <g>
+      <g id="mask">
+        <path className="masking__layer" d="M676.7373,352.80078a6.03487,6.03487,0,0,0-4.04394,1.376,6.20645,6.20645,0,0,0-2.06348,3.74121h12.21485a6.22125,6.22125,0,0,0-2.06348-3.74121A6.0402,6.0402,0,0,0,676.7373,352.80078Z" />
+        <path className="masking__layer" d="M534.78223,366.77637a3.0868,3.0868,0,0,0,1.293,2.61426,5.82291,5.82291,0,0,0,3.54882.96289,7.46452,7.46452,0,0,0,3.90625-1.01856,5.56643,5.56643,0,0,0,2.4209-2.99805v-3.02636h-5.88672Q534.78223,363.31055,534.78223,366.77637Z" />
+        <path className="masking__layer" d="M340.33594,352.80078a6.03489,6.03489,0,0,0-4.04395,1.376,6.20639,6.20639,0,0,0-2.06347,3.74121h12.21484a6.22119,6.22119,0,0,0-2.06348-3.74121A6.04019,6.04019,0,0,0,340.33594,352.80078Z" />
+        <path className="masking__layer" d="M430.07617,352.80078a6.03487,6.03487,0,0,0-4.04394,1.376,6.20645,6.20645,0,0,0-2.06348,3.74121h12.21484a6.22118,6.22118,0,0,0-2.06347-3.74121A6.04022,6.04022,0,0,0,430.07617,352.80078Z" />
+        <path className="masking__layer" d="M793.49414,352.80078a6.03487,6.03487,0,0,0-4.04394,1.376,6.20645,6.20645,0,0,0-2.06348,3.74121h12.21484a6.22118,6.22118,0,0,0-2.06347-3.74121A6.04023,6.04023,0,0,0,793.49414,352.80078Z" />
+        <path className="masking__layer" d="M758.16992,352.80078a6.03487,6.03487,0,0,0-4.04394,1.376,6.20645,6.20645,0,0,0-2.06348,3.74121h12.21484a6.22118,6.22118,0,0,0-2.06347-3.74121A6.04022,6.04022,0,0,0,758.16992,352.80078Z" />
+        <path className="masking__layer" d="M162.83594,352.14063a7.39054,7.39054,0,0,0-4.89746,1.65136,7.02385,7.02385,0,0,0-2.4209,4.51172h14.58105a7.10047,7.10047,0,0,0-2.4209-4.48437A7.236,7.236,0,0,0,162.83594,352.14063Z" />
+        <path className="masking__layer" d="M-.67383-138.11322V558.002H900V-138.11322Zm107.333,482.55072H94.334v31.25293H85.41992V344.4375H73.09521v-7.2627h33.564Zm34.44336,31.25293h-8.583v-15.626a7.72077,7.72077,0,0,0-1.541-5.25488,5.65139,5.65139,0,0,0-4.457-1.7334,6.83138,6.83138,0,0,0-5.17187,2.00879q-1.92627,2.0083-1.92578,5.96973v14.63574h-8.583V334.86426h8.583v14.30566a11.4233,11.4233,0,0,1,4.127-2.61426,14.94,14.94,0,0,1,5.28223-.90722,12.18535,12.18535,0,0,1,8.88574,3.30078q3.38379,3.30176,3.38379,9.79394Zm36.91894-12.37988H155.62793a6.96689,6.96689,0,0,0,2.86133,4.34668,9.46836,9.46836,0,0,0,5.6123,1.5957,11.35478,11.35478,0,0,0,4.09864-.6875,9.95055,9.95055,0,0,0,3.3291-2.17383l4.5664,4.95215q-4.18212,4.78712-12.21484,4.78711a19.32693,19.32693,0,0,1-8.8584-1.95313,14.31276,14.31276,0,0,1-5.94238-5.41992,14.92847,14.92847,0,0,1-2.09082-7.86816,15.1268,15.1268,0,0,1,2.06347-7.84082,14.45764,14.45764,0,0,1,5.667-5.44727,17.54888,17.54888,0,0,1,15.92871-.083,13.66966,13.66966,0,0,1,5.53027,5.36524,15.99735,15.99735,0,0,1,2.00781,8.11523Q178.18652,361.16456,178.02148,363.31055Zm52.21582,12.37988h-9.959V372.499a10.70577,10.70577,0,0,1-3.96094,2.751,12.92426,12.92426,0,0,1-4.89746.93555,12.50841,12.50841,0,0,1-9.21582-3.41114q-3.43945-3.41161-3.43945-10.23437V345.86816h10.4541V360.834q0,3.35744,1.26563,4.89747a4.49446,4.49446,0,0,0,3.68652,1.541,5.20322,5.20322,0,0,0,4.04394-1.7334q1.56885-1.73291,1.56836-5.30957V345.86816h10.4541Zm38.46,0H258.24316V360.33887q0-6.10693-4.89746-6.10742a5.59536,5.59536,0,0,0-4.31933,1.76074,7.56078,7.56078,0,0,0-1.623,5.28222v14.416h-10.4541V345.86816h9.959v3.2461a11.8628,11.8628,0,0,1,4.18164-2.77832,14.16753,14.16753,0,0,1,5.28223-.96289,12.25655,12.25655,0,0,1,8.94141,3.30175q3.38379,3.3003,3.38379,9.959Zm17.002,0h-10.4541V345.86816h10.4541Zm-.60547-34.74707a7.43883,7.43883,0,0,1-9.24316.02832,5.33779,5.33779,0,0,1,0-7.92383,6.61607,6.61607,0,0,1,4.62109-1.5957,6.84393,6.84393,0,0,1,4.62207,1.51367,4.88366,4.88366,0,0,1,1.76074,3.87891A5.3335,5.3335,0,0,1,285.09375,340.94336Zm15.68066,34.74707-12.2705-29.82227H299.2334l7.15332,18.21289,7.53808-18.21289h9.959l-12.3252,29.82227Zm55.29786-12.26953H334.39355a6.06556,6.06556,0,0,0,2.5586,3.54883,8.54207,8.54207,0,0,0,4.75879,1.2373,10.30466,10.30466,0,0,0,3.71386-.60449,10.941,10.941,0,0,0,3.16407-1.98145l5.502,5.72266q-4.34619,4.84278-12.71,4.8418a20.29063,20.29063,0,0,1-9.18848-1.98047,15.05656,15.05656,0,0,1-.33008-26.85156,18.77474,18.77474,0,0,1,16.50684-.13672,14.0273,14.0273,0,0,1,5.75,5.33691,15.67768,15.67768,0,0,1,2.11816,8.28027Q356.2373,360.94531,356.07227,363.4209Zm24.97949-8.63867a21.29983,21.29983,0,0,0-2.36524-.165,7.28812,7.28812,0,0,0-5.25488,1.81543q-1.89843,1.8164-1.89844,5.55761v13.7002H361.0791V345.86816h9.959v3.57618a9.74136,9.74136,0,0,1,4.15429-3.05371,15.77482,15.77482,0,0,1,5.85938-1.01758Zm28.22656,16.64355a11.10107,11.10107,0,0,1-5.03516,3.49414,22.56459,22.56459,0,0,1-8.00586,1.26563,32.14535,32.14535,0,0,1-7.37207-.85254,19.46029,19.46029,0,0,1-5.833-2.22852l3.13672-7.15332a17.6654,17.6654,0,0,0,4.84179,2.00879,21.65051,21.65051,0,0,0,5.55762.74219,8.8812,8.8812,0,0,0,3.60352-.52246,1.63935,1.63935,0,0,0,1.07324-1.5127q0-.99022-1.293-1.40332a26.74412,26.74412,0,0,0-4.09961-.79785,40.45473,40.45473,0,0,1-6.08008-1.15527,9.37422,9.37422,0,0,1-4.31933-2.6416,7.55224,7.55224,0,0,1-1.81543-5.39161,8.14751,8.14751,0,0,1,1.70605-5.08984,11.30163,11.30163,0,0,1,4.97949-3.52148,21.342,21.342,0,0,1,7.83985-1.293,31.68683,31.68683,0,0,1,6.43847.66015,18.52615,18.52615,0,0,1,5.33692,1.8711l-3.13672,7.09765a17.38929,17.38929,0,0,0-8.583-2.20117,8.564,8.564,0,0,0-3.63184.57813,1.72125,1.72125,0,0,0-1.15527,1.51269q0,1.04591,1.26563,1.458a28.85622,28.85622,0,0,0,4.18164.85352,46.24451,46.24451,0,0,1,6.10742,1.26562,9.33494,9.33494,0,0,1,4.209,2.61328,7.44451,7.44451,0,0,1,1.78809,5.30957A8.06982,8.06982,0,0,1,409.27832,371.42578Zm1.3811-125.01532a73.69524,73.69524,0,0,1-9.02288,9.119,97.05608,97.05608,0,0,1-9.023,6.95868,96.06841,96.06841,0,0,1-18.04615,9.6369,93.9412,93.9412,0,0,1-36.092,6.83649,96.6814,96.6814,0,0,1-36.092-7.08991,97.96978,97.96978,0,0,1-27.069-16.47827,74.32612,74.32612,0,0,1-9.023-8.98291,7.90874,7.90874,0,0,1,0-10.14,74.32946,74.32946,0,0,1,9.023-8.98279,97.973,97.973,0,0,1,27.069-16.47839,96.6816,96.6816,0,0,1,36.092-7.08985,93.9377,93.9377,0,0,1,36.092,6.83655,96.06838,96.06838,0,0,1,18.04615,9.63684,97.0474,97.0474,0,0,1,9.02288,6.95868,73.686,73.686,0,0,1,9.023,9.119A7.98911,7.98911,0,0,1,410.65942,246.41046ZM445.8125,363.4209H424.13379a6.06554,6.06554,0,0,0,2.55859,3.54883,8.54211,8.54211,0,0,0,4.75879,1.2373,10.30468,10.30468,0,0,0,3.71387-.60449,10.94087,10.94087,0,0,0,3.16406-1.98145l5.502,5.72266q-4.34619,4.84278-12.71,4.8418a20.29055,20.29055,0,0,1-9.18847-1.98047,15.05656,15.05656,0,0,1-.33008-26.85156,18.77474,18.77474,0,0,1,16.50684-.13672,14.02736,14.02736,0,0,1,5.75,5.33691,15.67768,15.67768,0,0,1,2.11816,8.28027Q445.97754,360.94531,445.8125,363.4209Zm9.35718-178.64057a7.989,7.989,0,0,1-10.14,0,73.68537,73.68537,0,0,1-9.11895-9.02294,97.04714,97.04714,0,0,1-6.95868-9.02289,96.07024,96.07024,0,0,1-9.63684-18.04615,93.93991,93.93991,0,0,1-6.83655-36.092,96.68178,96.68178,0,0,1,7.08984-36.092,97.97244,97.97244,0,0,1,16.4784-27.069,74.32659,74.32659,0,0,1,8.98291-9.023,7.90872,7.90872,0,0,1,10.13989,0,74.33236,74.33236,0,0,1,8.98291,9.023,97.971,97.971,0,0,1,16.47833,27.069,96.68225,96.68225,0,0,1,7.08984,36.092,93.94115,93.94115,0,0,1-6.83648,36.092,96.07084,96.07084,0,0,1-9.6369,18.04615,97.05608,97.05608,0,0,1-6.95868,9.023A73.69406,73.69406,0,0,1,455.16968,184.78033Zm20.40942,190.9101h-8.583V346.08887h8.583Zm-.44043-35.13184a5.38053,5.38053,0,0,1-3.85156,1.40332,5.44184,5.44184,0,0,1-3.85156-1.376,4.65954,4.65954,0,0,1,0-6.82227,5.4462,5.4462,0,0,1,3.85156-1.376,5.58279,5.58279,0,0,1,3.85156,1.32031,4.234,4.234,0,0,1,1.48633,3.30176A4.66985,4.66985,0,0,1,475.13867,340.55859Zm30.70313,31.06055a10.53632,10.53632,0,0,1-4.81446,3.32813,21.44662,21.44662,0,0,1-7.51074,1.18359,28.73891,28.73891,0,0,1-7.208-.9082,18.83512,18.83512,0,0,1-5.6123-2.28321l2.86133-6.16308a17.94672,17.94672,0,0,0,4.78711,2.06347,20.04183,20.04183,0,0,0,5.50195.79786q5.44629,0,5.44727-2.69629a1.9087,1.9087,0,0,0-1.48633-1.81543,22.111,22.111,0,0,0-4.56641-.93555,43.81862,43.81862,0,0,1-5.99707-1.26563,9.29227,9.29227,0,0,1-4.09961-2.53027,7.19927,7.19927,0,0,1-1.7334-5.17285,8.10624,8.10624,0,0,1,1.623-4.97949,10.45336,10.45336,0,0,1,4.73242-3.38379,20.23616,20.23616,0,0,1,7.3457-1.21,28.86592,28.86592,0,0,1,6.24512.6875,17.49335,17.49335,0,0,1,5.14453,1.89844l-2.86132,6.10742a17.08891,17.08891,0,0,0-8.52833-2.20117,8.67143,8.67143,0,0,0-4.127.7705,2.26747,2.26747,0,0,0-1.376,1.98047,1.965,1.965,0,0,0,1.48632,1.92578,26.22545,26.22545,0,0,0,4.73145,1.0459,50.89912,50.89912,0,0,1,5.94238,1.293,8.929,8.929,0,0,1,4.0166,2.50391,7.10392,7.10392,0,0,1,1.70606,5.06152A7.82138,7.82138,0,0,1,505.8418,371.61914Zm48.69336,4.07129H546.502v-3.68652q-2.42138,4.12645-9.02343,4.12695a14.0138,14.0138,0,0,1-5.915-1.15527,8.81257,8.81257,0,0,1-3.82422-3.19141,8.30152,8.30152,0,0,1-1.32031-4.62207,7.7153,7.7153,0,0,1,3.10839-6.49316q3.10841-2.36427,9.60157-2.36524h6.82226a5.471,5.471,0,0,0-1.70508-4.31933,7.51519,7.51519,0,0,0-5.11718-1.51368,14.93649,14.93649,0,0,0-4.64942.74317,12.31065,12.31065,0,0,0-3.8789,2.00781l-3.08106-5.99707a18.60224,18.60224,0,0,1,5.80469-2.64063,25.9807,25.9807,0,0,1,6.96-.93554q6.87744,0,10.56445,3.27344,3.68555,3.27391,3.68653,9.877Zm16.34082,0h-8.583V334.86426h8.583Zm16.56152,0h-8.583V334.86426h8.583Zm-25.71344-96.72888a93.94068,93.94068,0,0,1-36.092-6.83649,96.06851,96.06851,0,0,1-18.0462-9.6369,97.05262,97.05262,0,0,1-9.02295-6.95868,73.69532,73.69532,0,0,1-9.02289-9.119,7.98911,7.98911,0,0,1,0-10.14,73.69043,73.69043,0,0,1,9.023-9.119,97.05417,97.05417,0,0,1,9.02289-6.95868,96.06849,96.06849,0,0,1,18.0462-9.63684,93.93718,93.93718,0,0,1,36.092-6.83655,96.68105,96.68105,0,0,1,36.092,7.08985,97.97217,97.97217,0,0,1,27.069,16.47839,74.32923,74.32923,0,0,1,9.023,8.98279,7.90888,7.90888,0,0,1,0,10.14,74.3259,74.3259,0,0,1-9.023,8.98291,97.9689,97.9689,0,0,1-27.069,16.47827A96.68094,96.68094,0,0,1,561.72406,278.96155Zm87.44782,96.72888h-10.124l-5.6123-16.23145-5.833,16.23145h-10.124l-10.61914-29.82227h9.90429l6.10742,18.04786,6.32715-18.04786h8.91406l6.16211,18.21289,6.32813-18.21289H659.791Zm43.30175-12.26953H670.79492a6.06556,6.06556,0,0,0,2.5586,3.54883,8.54206,8.54206,0,0,0,4.75878,1.2373,10.30468,10.30468,0,0,0,3.71387-.60449,10.94087,10.94087,0,0,0,3.16406-1.98145l5.502,5.72266q-4.34619,4.84278-12.71,4.8418a20.29063,20.29063,0,0,1-9.18848-1.98047,15.05656,15.05656,0,0,1-.33008-26.85156,18.77474,18.77474,0,0,1,16.50684-.13672,14.0273,14.0273,0,0,1,5.75,5.33691,15.67768,15.67768,0,0,1,2.11816,8.28027Q692.63867,360.94531,692.47363,363.4209Zm44.89746,8.00488a11.10109,11.10109,0,0,1-5.03515,3.49414,22.56467,22.56467,0,0,1-8.00586,1.26563,32.14535,32.14535,0,0,1-7.37207-.85254,19.46036,19.46036,0,0,1-5.833-2.22852l3.13672-7.15332a17.66534,17.66534,0,0,0,4.8418,2.00879,21.65038,21.65038,0,0,0,5.55761.74219,8.88116,8.88116,0,0,0,3.60352-.52246,1.63935,1.63935,0,0,0,1.07324-1.5127q0-.99022-1.293-1.40332a26.7436,26.7436,0,0,0-4.09961-.79785,40.455,40.455,0,0,1-6.08008-1.15527,9.37427,9.37427,0,0,1-4.31933-2.6416,7.55228,7.55228,0,0,1-1.81543-5.39161,8.14757,8.14757,0,0,1,1.70605-5.08984,11.30166,11.30166,0,0,1,4.9795-3.52148,21.342,21.342,0,0,1,7.83984-1.293,31.68693,31.68693,0,0,1,6.43848.66015,18.52608,18.52608,0,0,1,5.33691,1.8711l-3.13672,7.09765a17.38923,17.38923,0,0,0-8.583-2.20117,8.56393,8.56393,0,0,0-3.63183.57813,1.72124,1.72124,0,0,0-1.15528,1.51269q0,1.04591,1.26563,1.458a28.85622,28.85622,0,0,0,4.18164.85352,46.24451,46.24451,0,0,1,6.10742,1.26562,9.335,9.335,0,0,1,4.209,2.61328,7.44455,7.44455,0,0,1,1.78808,5.30957A8.06982,8.06982,0,0,1,737.37109,371.42578Zm36.53516-8.00488H752.22754a6.06554,6.06554,0,0,0,2.55859,3.54883,8.54211,8.54211,0,0,0,4.75879,1.2373,10.30468,10.30468,0,0,0,3.71387-.60449,10.94087,10.94087,0,0,0,3.16406-1.98145l5.50195,5.72266q-4.34619,4.84278-12.71,4.8418a20.29055,20.29055,0,0,1-9.18847-1.98047,15.05656,15.05656,0,0,1-.33008-26.85156,18.77474,18.77474,0,0,1,16.50684-.13672,14.02736,14.02736,0,0,1,5.75,5.33691,15.67768,15.67768,0,0,1,2.11816,8.28027Q774.07129,360.94531,773.90625,363.4209Zm35.32422,0H787.55176a6.06554,6.06554,0,0,0,2.55859,3.54883,8.54209,8.54209,0,0,0,4.75879,1.2373,10.30471,10.30471,0,0,0,3.71387-.60449,10.94087,10.94087,0,0,0,3.16406-1.98145l5.50195,5.72266q-4.34619,4.84278-12.71,4.8418a20.29055,20.29055,0,0,1-9.18847-1.98047,15.05656,15.05656,0,0,1-.33008-26.85156,18.77472,18.77472,0,0,1,16.50683-.13672,14.02724,14.02724,0,0,1,5.75,5.33691,15.67759,15.67759,0,0,1,2.11817,8.28027Q809.39551,360.94531,809.23047,363.4209Zm13.89258,11.0039a6.55485,6.55485,0,0,1-8.94043,0,5.94309,5.94309,0,0,1-1.81641-4.457,5.8769,5.8769,0,0,1,1.78809-4.456,6.78256,6.78256,0,0,1,8.99609,0,5.92524,5.92524,0,0,1,1.76074,4.456A5.99363,5.99363,0,0,1,823.12305,374.4248Z" />
+      </g>
+
+      <g id="visible">
+        <polygon className="visible__layer" points="85.12 375.99 85.12 344.737 72.795 344.737 72.795 336.875 106.959 336.875 106.959 344.737 94.634 344.737 94.634 375.99 85.12 375.99" />
+        <path className="visible__layer" d="M132.21973,375.99023V360.06445a7.47226,7.47226,0,0,0-1.46582-5.05566,5.38977,5.38977,0,0,0-4.23243-1.63281,6.56969,6.56969,0,0,0-4.95507,1.917c-1.22266,1.27441-1.84278,3.21289-1.84278,5.76172v14.93554H110.541V334.56445h9.18261v13.91309a11.94548,11.94548,0,0,1,3.72168-2.20313,15.31784,15.31784,0,0,1,5.3877-.92578,12.52175,12.52175,0,0,1,9.09472,3.38575c2.30567,2.25,3.47461,5.61718,3.47461,10.00878v17.24707Z" />
+        <path className="visible__layer" d="M163.88086,376.43066a19.71374,19.71374,0,0,1-8.99414-1.98535,14.6779,14.6779,0,0,1-6.06348-5.53222,15.29624,15.29624,0,0,1-2.13379-8.02344,15.49428,15.49428,0,0,1,2.1045-7.99317,14.814,14.814,0,0,1,5.7832-5.55859,17.84722,17.84722,0,0,1,16.21191-.084,14.02525,14.02525,0,0,1,5.64942,5.48047,16.36122,16.36122,0,0,1,2.04785,8.26464c0,.04493,0,.1836-.166,2.335l-.02148.27637H156.00977a6.67438,6.67438,0,0,0,2.65234,3.80176,9.21855,9.21855,0,0,0,5.43945,1.541,11.10818,11.10818,0,0,0,3.99121-.668,9.70064,9.70064,0,0,0,3.22852-2.10938l.2207-.21289,4.957,5.375-.17773.20313C173.4873,374.78613,169.30176,376.43066,163.88086,376.43066Zm5.86035-18.42675a6.75814,6.75814,0,0,0-2.25781-3.95606,6.97379,6.97379,0,0,0-4.64746-1.60742,7.12971,7.12971,0,0,0-4.70508,1.582,6.65946,6.65946,0,0,0-2.25684,3.98145Z" />
+        <path className="visible__layer" d="M471.28711,342.26172a5.75488,5.75488,0,0,1-4.05566-1.45606,4.9596,4.9596,0,0,1,0-7.26171,5.75985,5.75985,0,0,1,4.05566-1.45606,5.89937,5.89937,0,0,1,4.05078,1.39551,4.54292,4.54292,0,0,1,1.58691,3.52637,4.97757,4.97757,0,0,1-1.58007,3.7666A5.69366,5.69366,0,0,1,471.28711,342.26172Z" />
+        <rect className="visible__layer" x="466.69629" y="345.78906" width="9.18262" height="30.20117" />
+        <path className="visible__layer" d="M493.5166,376.43066a29.15165,29.15165,0,0,1-7.2832-.918,19.20961,19.20961,0,0,1-5.70215-2.32227l-.21582-.14258,3.11426-6.709.29.18457a17.71625,17.71625,0,0,0,4.707,2.02734,19.816,19.816,0,0,0,5.41992.78711c3.415,0,5.14746-.80664,5.14746-2.39648,0-.71094-.42187-1.21289-1.291-1.53418a21.98352,21.98352,0,0,0-4.499-.919,44.29056,44.29056,0,0,1-6.04687-1.27637,9.62493,9.62493,0,0,1-4.2295-2.61035,7.5184,7.5184,0,0,1-1.8164-5.37988,8.4354,8.4354,0,0,1,1.68262-5.15918,10.79157,10.79157,0,0,1,4.86328-3.4834,20.63078,20.63078,0,0,1,7.45507-1.23047,29.29763,29.29763,0,0,1,6.30958.69434,17.84539,17.84539,0,0,1,5.23339,1.93359l.2334.13867-3.11328,6.64649-.28223-.1582a16.86425,16.86425,0,0,0-8.38086-2.16309,8.433,8.433,0,0,0-3.98046.73242,1.97741,1.97741,0,0,0-1.22266,1.71875,1.67879,1.67879,0,0,0,1.29,1.64453,26.15445,26.15445,0,0,0,4.67285,1.03028,51.42909,51.42909,0,0,1,5.9834,1.30273,9.26377,9.26377,0,0,1,4.14941,2.58594,7.42076,7.42076,0,0,1,1.78711,5.2666,8.15306,8.15306,0,0,1-1.71191,5.08008,10.88453,10.88453,0,0,1-4.94727,3.42676A21.85367,21.85367,0,0,1,493.5166,376.43066Z" />
+        <path className="visible__layer" d="M537.47852,376.43066a14.3744,14.3744,0,0,1-6.041-1.18261,9.15088,9.15088,0,0,1-3.9502-3.30078,8.63549,8.63549,0,0,1-1.36816-4.78516,8.03169,8.03169,0,0,1,3.22656-6.73145c2.11719-1.61035,5.40821-2.42675,9.78321-2.42675h6.5166a5.08106,5.08106,0,0,0-1.59863-3.79493,7.26752,7.26752,0,0,0-4.918-1.43847,14.706,14.706,0,0,0-4.55664.72851,12.0561,12.0561,0,0,0-3.78516,1.958l-.28613.22656-3.36719-6.55371.21289-.15039a18.98933,18.98933,0,0,1,5.89746-2.68457,26.409,26.409,0,0,1,7.04-.94629c4.63867,0,8.25977,1.127,10.76367,3.34864,2.5127,2.23242,3.78711,5.63086,3.78711,10.10156v17.1914h-8.63281v-3.00488C544.44043,375.27246,541.51172,376.43066,537.47852,376.43066Zm2.58593-12.82031c-3.35254,0-4.98242,1.03613-4.98242,3.166a2.78012,2.78012,0,0,0,1.17188,2.374,5.567,5.567,0,0,0,3.37011.90332,7.19441,7.19441,0,0,0,3.752-.97656,5.28457,5.28457,0,0,0,2.27539-2.793v-2.67383Z" />
+        <rect className="visible__layer" x="561.99316" y="334.56445" width="9.18262" height="41.42578" />
+        <rect className="visible__layer" x="578.55469" y="334.56445" width="9.18262" height="41.42578" />
+      </g>
+    </g>
+  )
+}
+
+export default En
